% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{apply_filter_predicate}
\alias{apply_filter_predicate}
\title{Intermediate function to apply filter predicates on a data.frame}
\usage{
apply_filter_predicate(df, verbose, ...)
}
\arguments{
\item{df}{data.frame we want to apply filter(s) expression(s)}

\item{verbose}{Show (\code{TRUE}) or not (\code{FALSE}) the filter predicate expression}

\item{...}{filter predicates to apply to \code{df}}
}
\value{
a data.frame
}
\description{
This function is used internally by all the \verb{get_*()} functions to filter on
deployments.
}
\examples{
# and
apply_filter_predicate(mica$deployments,
                       verbose = TRUE,
                       pred_gte("latitude", 51.28),
                       pred_lt("longitude", 3.56))
# equivalent of
apply_filter_predicate(mica$deployments,
                       verbose = TRUE,
                       pred_and(pred_gte("latitude", 51.28),
                                pred_lt("longitude", 3.56)))


# or
apply_filter_predicate(mica$deployments,
                       verbose = TRUE,
                       pred_or(pred_gte("latitude", 51.28),
                                pred_lt("longitude", 3.56)))
}
