% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_individual_radius_angle.R
\name{calculate_individual_radius_angle}
\alias{calculate_individual_radius_angle}
\title{Calculate animal position}
\usage{
calculate_individual_radius_angle(
  animal_positions,
  calibration_models,
  deployment_id = "deploymentID",
  event_id = "eventID",
  x = "x",
  y = "y",
  image_width = "imageWidth",
  image_height = "imageHeight"
)
}
\arguments{
\item{animal_positions}{Data frame (tibble) of animal position digitization
data. It must contain (at least) the columns defined in args
\code{deployment_id}, \code{event_id}, \code{x}, \code{y}, \code{image_width} and \code{image_height}.}

\item{calibration_models}{Named list of deployment calibration models or site
calibration models (\code{calibs} objects), produced using \code{cal.site()} (not yet
included in this package). The deployment names are used as names.}

\item{deployment_id}{Column in \code{animal_positions} against which names of the
elements can be matched to apply the right deployment calibration models.
Default: \code{"deploymentID"}.}

\item{event_id}{Column in \code{animal_positions} containing the event ID the
images belong to. Default: \code{"eventID"}.}

\item{x}{Column in \code{animal_positions} containing x pixel positions for each
digitised point. Default: \code{"x"}.}

\item{y}{Column in \code{animal_positions} containing y pixel positions for each
digitised point. Default: \code{"y"}.}

\item{image_width}{Column in \code{animal_positions} containing the pixel x
dimension of each image. Default: \code{"imageWidth"}. Notice that the pixel x
dimension must be consistent for each deployment.}

\item{image_height}{Column in \code{animal_positions} containing the pixel y
dimension of each image. Default: \code{"imageHeight"}. Notice that the pixel y
dimension must be consistent for each deployment.}
}
\value{
Original tibble data frame as passed via \code{animal_positions} with
additional columns:
\itemize{
\item \code{radius}: Radial distance from camera.
\item \code{angle}: Angular distance from camera.
\item \code{frame_count}: Indicator of the frame order within each sequence.
}
}
\description{
Calculates the position of animal relative to a camera based on image pixel
positions and site calibration models.
}
\examples{
# Use default values
calculate_individual_radius_angle(animal_positions, calibration_models)
}
\concept{density estimation functions}
