% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camtrapdp}
\alias{camtrapdp}
\title{Sample of camtrap-dp formatted data}
\format{
An object of class \code{list} of length 4.
}
\source{
https://github.com/tdwg/dwc-for-biologging/tree/403f57db105982dc05b70f3cf66fd2b5591798db/derived/camtrap-dp/data/raw
}
\usage{
camtrapdp
}
\description{
A sample camera trap data package of musk rat and coypu from MICA project.
}
\keyword{datasets}
