% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapR_cameraOperation.R
\name{camtrapR_cameraOperation}
\alias{camtrapR_cameraOperation}
\title{Get camera operation matrix}
\usage{
camtrapR_cameraOperation(
  x,
  station_col = "locationName",
  camera_col = NULL,
  session_col = NULL,
  use_prefix = FALSE
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{station_col}{Column name to use for identifying the stations. Default:
\code{"locationName"}.}

\item{camera_col}{Column name of the column specifying Camera ID. Default:
\code{NULL}.}

\item{session_col}{Column name to use for identifying the session. Default:
\code{NULL}. Use it for creating multi-session / multi-season detection
histories.}

\item{use_prefix}{Logical (\code{TRUE} or \code{FALSE}). If \code{TRUE} the returned row
names will start with prefix \code{"Station"} as returned by
\href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camtrapR::cameraOperation()}.
Default: \code{FALSE}.}
}
\value{
A matrix. Row names always indicate the station ID. Column names are
dates.
}
\description{
Returns the \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camera operation matrix} as
returned by \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camtrapR::cameraOperation()}.
}
\details{
The deployment data are by default grouped by \code{locationName} (station ID in
camtrapR jargon) or another column specified by the user via the
\code{station_col} argument. If multiple deployments are linked to same location,
daily efforts higher than 1 occur.

Partially active days, e.g. the first or the last day of a deployment, result
in decimal effort values as in \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camtrapR::cameraOperation()}.
}
\examples{
library(dplyr)
library(stringr)

x <- example_dataset()
camtrapR_cameraOperation(x)

# Specify column with station names
camtrapR_cameraOperation(x, station_col = "locationID")

# Specify column with session IDs
x_sessions <- x
deployments(x_sessions) <- deployments(x_sessions) \%>\%
  mutate(session = ifelse(
    str_starts(.data$locationName, "B_DL_"),
    "after2020",
    "before2020"
  ))
  
camtrapR_cameraOperation(x_sessions, session_col = "session")

# Specify column with camera IDs
x_cameras <- x_sessions
deployments(x_cameras) <- deployments(x_cameras) \%>\%
  mutate(cameraID = c(1, 2, 3, 4))
camtrapR_cameraOperation(x_cameras, camera_col = "cameraID")

# Specify both session and camera IDs
camtrapR_cameraOperation(
  x_cameras,
  camera_col = "cameraID",
  session_col = "session"
)

# Use prefix Station as in camtrapR's camera operation matrix
camtrapR_cameraOperation(x, use_prefix = TRUE)
}
\seealso{
Other camtrapR-derived functions:
\code{\link[=camtrapR_detectionHistory]{camtrapR_detectionHistory()}},
\code{\link[=camtrapR_recordTable]{camtrapR_recordTable()}}
}
\concept{camtrapR-derived functions}
