% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R
\name{camtraptor-deprecated}
\alias{camtraptor-deprecated}
\alias{mica}
\alias{animal_pos}
\alias{calib_models}
\title{Deprecated datasets in camtraptor}
\description{
\itemize{
\item The \code{mica} dataset was a Camtrap DP 0.1.6 data package and it is not
supported anymore. Replaced by \code{\link[=example_dataset]{example_dataset()}}.
\item The sample of animal position digitization data \code{animal_pos} is deprecated.
Use \code{animal_positions} instead.
\item The sample of deployment calibration models \code{calib_models} is deprecated.
Use \code{calibration_models} instead.
}
}
\seealso{
Other sample data:
\code{\link{animal_positions}},
\code{\link{calibration_models}}
}
\concept{sample data}
