% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{check_filter_arg}
\alias{check_filter_arg}
\title{Check filter argument}
\usage{
check_filter_arg(arg)
}
\arguments{
\item{arg}{character with the argument name of the filter predicate}
}
\value{
\code{TRUE} or an error message
}
\description{
Check that the filter argument in a filter predicate is a character and has
length one.
}
\examples{
\dontrun{
check_filter_arg("latitude")
check_filter_arg("location_name")
# this returns an error
check_filter_arg(5)
}
}
\keyword{internal}
