% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{check_filter_symbol}
\alias{check_filter_symbol}
\title{Check filter symbol}
\usage{
check_filter_symbol(symbol)
}
\arguments{
\item{symbol}{Character with symbol for filter predicate, e.g. "==".}
}
\value{
\code{TRUE} or an error message.
}
\description{
Check that the symbol used in a filter predicate is a character and has
length one.
}
\examples{
\dontrun{
check_filter_symbol("==")
check_filter_symbol("!=")
# error: not a character
check_filter_symbol(5)
# error: length > 1
check_filter_symbol(c("==", "\%in\%"))
}
}
\keyword{internal}
