% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{check_filter_value_length}
\alias{check_filter_value_length}
\title{Check filter value length}
\usage{
check_filter_value_length(value)
}
\value{
\code{TRUE} or an error message
}
\description{
Check that the value in filter predicates has length one. Required for
basic filter predicates. Used in \code{check_filter_value()}.
}
\examples{
\dontrun{
check_filter_value_length(5)
check_filter_value_length("a")
# this returns an error
check_filter_value_length(c("a", "aa"))
}
}
