% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{check_filter_value_type}
\alias{check_filter_value_type}
\title{Check filter value type}
\usage{
check_filter_value_type(value)
}
\arguments{
\item{value}{Character, number, Date or POSIXct object}
}
\value{
\code{TRUE} or an error message
}
\description{
Check that the value argument in a filter predicate is one of the supported
types. Required for basic filter predicates.  Used in \code{check_filter_value()}.
}
\examples{
\dontrun{
check_filter_value_type("a")
check_filter_value_type(5)
# this returns an error
check_filter_value_type(list(5))
}
}
\keyword{internal}
