% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_summary}
\alias{check_summary}
\title{Check a data frame is output of \code{summarize_observations()}.}
\usage{
check_summary(grouped_df)
}
\arguments{
\item{grouped_df}{A grouped tibble data frame as returned by
\code{summarize_observations()}.}
}
\value{
If no error, \code{TRUE}.
}
\description{
Check a data frame is output of \code{summarize_observations()}.
}
