% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dep_calib_models}
\alias{dep_calib_models}
\title{Sample of deployment calibration models or site calibration models (\code{depcal} objects)}
\format{
An object of class \code{calibs} of length 4.
}
\usage{
dep_calib_models
}
\description{
A \code{calibs} object (list). The deployment names are used as names.
}
\seealso{
Other sample data: 
\code{\link{animal_positions}},
\code{\link{mica}}
}
\concept{sample data}
\keyword{datasets}
