% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{deployments}
\alias{deployments}
\alias{deployments<-}
\title{Get or set deployments}
\usage{
deployments(x)

deployments(x) <- value
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}

\item{value}{A data frame to assign as deployments.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with deployments.
}
\description{
Re-export of \code{\link[camtrapdp:deployments]{camtrapdp::deployments()}}.
}
\examples{
x <- example_dataset()
# Get deployments
deployments(x)

# Set deployments (not recommended outside a function)
deployments(x) <- head(deployments(x), 1)
}
\seealso{
Other accessor functions: 
\code{\link{add_coordinates}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{taxa}()}
}
\concept{accessor functions}
