% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{.features_deployments}
\alias{.features_deployments}
\title{Features returned by \code{summarize_deployments()}. Useful to check output of
the function and other check functions.}
\usage{
.features_deployments
}
\description{
Features returned by \code{summarize_deployments()}. Useful to check output of
the function and other check functions.
}
\keyword{internal}
