% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{.features_observations}
\alias{.features_observations}
\title{Features returned by \code{summarize_observations()}. Useful to check output of
the function and other check functions.}
\usage{
.features_observations
}
\description{
Features returned by \code{summarize_observations()}. Useful to check output of
the function and other check functions.
}
\keyword{internal}
