% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{.group_bys_deployments}
\alias{.group_bys_deployments}
\title{Define internal vectors with possible \code{group_by} values, to be used in
\code{summarize_observations()}, \code{summarize_deployments()} and check functions.}
\usage{
.group_bys_deployments
}
\description{
Define internal vectors with possible \code{group_by} values, to be used in
\code{summarize_observations()}, \code{summarize_deployments()} and check functions.
}
\keyword{internal}
