% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{.prefixes_for_hover_info}
\alias{.prefixes_for_hover_info}
\title{Prefix table for hover info over deployments}
\usage{
.prefixes_for_hover_info
}
\description{
Stores prefixes for info shown in leaflet map while hovering over a
deployment with the mouse.
}
\details{
A data.frame of all prefixes with the following columns:
\itemize{
\item \code{info}: all valid grouping columns (deployments or observations)
and features.
\item \code{prefix}: Prefix to use.
}
}
\keyword{internal}
