% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{example_dataset}
\alias{example_dataset}
\title{Read the Camtrap DP example dataset}
\usage{
example_dataset()
}
\value{
Camera Trap Data Package object.
}
\description{
Re-export of \code{\link[camtrapdp:example_dataset]{camtrapdp::example_dataset()}}.
}
\examples{
example_dataset()
}
\seealso{
Other read sample data functions:
\code{\link{mica}}
}
\concept{read sample data functions}
