% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_timelapse.R
\name{filter_out_timelapse}
\alias{filter_out_timelapse}
\title{Filter out timelapse observations}
\usage{
filter_out_timelapse(x)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
\code{x} filtered.
}
\description{
Subsets observations in a Camera Trap Data Package object, removing timelapse
observations, i.e. observations where \code{captureMethod} = \code{timeLapse}. This
function is a shortcut for \code{filter_observations(x, captureMethod != "timelapse")}.
}
\examples{
x <- example_dataset()

# `x` doesn't contain timelapse observations, returned as is
filter_out_timelapse(x)

# Create a data package with timelapse observations
obs <- observations(x)
obs$captureMethod <- c(rep("timelapse", nrow(obs) - 1), "activityDetection")
observations(x) <- obs
# Filter out timelapse observations
filter_out_timelapse(x)
}
\seealso{
Other filter functions:
\code{\link[=filter_deployments]{filter_deployments()}},
\code{\link[=filter_media]{filter_media()}},
\code{\link[=filter_observations]{filter_observations()}}
}
\concept{filter functions}
