% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_dep_no_obs}
\alias{get_dep_no_obs}
\title{Get deployments with no observations}
\usage{
get_dep_no_obs(deployments, observations)
}
\arguments{
\item{deployments}{a tibble (data.frame) containing deployments}

\item{observations}{a tibble (data.frame) containing observations}
}
\value{
a tibble (data.frame) with deployments not linked to any observations
}
\description{
Return subset of deployments without observations. A message is also returned
to list the ID of such deployments.
}
