% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{individuals}
\alias{individuals}
\title{Get individuals}
\usage{
individuals(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the individuals that have an
\code{individualID}, containing the following columns:
\itemize{
\item \code{individualID}
\item \code{scientificName}
\item \code{lifeStage}
\item \code{sex}
}
}
\description{
Re-export of \code{\link[camtrapdp:individuals]{camtrapdp::individuals()}}.
}
\examples{
x <- example_dataset()
individuals(x)
}
\seealso{
Other accessor functions: 
\code{\link{add_coordinates}()},
\code{\link{contributors}()},
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{taxa}()}
}
\concept{accessor functions}
