% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{locations}
\alias{locations}
\title{Get locations}
\usage{
locations(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the locations, containing the
following columns:
\itemize{
\item \code{locationID}
\item \code{locationName}
\item \code{latitude}
\item \code{longitude}
\item \code{coordinateUncertainty}
}
}
\description{
Re-export of \code{\link[camtrapdp:locations]{camtrapdp::locations()}}.
}
\examples{
x <- example_dataset()
locations(x)
}
\seealso{
Other accessor functions: 
\code{\link{add_coordinates}()},
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{taxa}()}
}
\concept{accessor functions}
