% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dep.R
\name{map_dep}
\alias{map_dep}
\title{Visualize deployments features}
\usage{
map_dep(
  x,
  feature,
  species = NULL,
  sex = NULL,
  life_stage = NULL,
  effort_unit = NULL,
  cluster = TRUE,
  hover_columns = c("n", "species", "deploymentID", "locationID", "locationName",
    "latitude", "longitude", "start", "end"),
  palette = "inferno",
  zero_values_show = TRUE,
  zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/000000/multiply.png",
  zero_values_icon_size = 10,
  na_values_show = TRUE,
  na_values_icon_url = "https://img.icons8.com/ios-glyphs/30/FA5252/multiply.png",
  na_values_icon_size = 10,
  relative_scale = TRUE,
  max_scale = NULL,
  radius_range = c(10, 50)
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{feature}{Feature to visualize. One of the columns generated by
\code{\link[=summarize_deployments]{summarize_deployments()}} or \code{\link[=summarize_observations]{summarize_observations()}} not used for
grouping. Possible values present in output of \code{\link[=summarize_observations]{summarize_observations()}}
are:
\itemize{
\item \code{n_scientificName}: Number of detected species.
\item \code{n_events}: Number of events observed.
\item \code{n_observations}: Number of observations.
\item \code{sum_count}: Number of individuals observed.
\item \code{rai_observations}: RAI calculated using the number
of observations (\code{n_observations}).
\item \code{rai_count}: RAI calculated using the number
of observed individuals (\code{sum_count}).
}

Possible values present in output of \code{\link[=summarize_deployments]{summarize_deployments()}} are:
\itemize{
\item \code{effort_duration}: Deployment effort duration.
}}

\item{species}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Character with scientific
names. Common names are not supported anymore as of camtraptor 1.0.0.
Please, check \code{filter_observations()} to know how to filter by
\code{scientificName}.
If \code{"all"} (default) all scientific names are automatically selected.
If \code{NULL} all observations of all species are taken into account.}

\item{sex}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Character defining the sex
class to filter on, e.g. \code{"female"} or \code{c("male", "unknown")}.
If \code{NULL} (default) all observations of all sex classes are taken into
account. Please, check
\code{filter_observations()} to know how to filter by \code{sex}.}

\item{life_stage}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Character vector
defining the life stage class to filter on,
e.g. \code{"adult"} or \code{c("subadult", "adult")}.
If \code{NULL} (default) all observations of all life stage classes are taken
into account. Please, check \code{filter_observations()} to know how to filter
by \code{lifeStage}.}

\item{effort_unit}{Time unit to use while visualizing deployment effort
duration. Ignored if \code{feature} is not \code{"effort_duration"}. One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
}

If  \code{NULL} (default), the effort is returned in hours.}

\item{cluster}{Logical value indicating whether using the cluster option
while visualizing maps. Default: \code{TRUE}.}

\item{hover_columns}{Character vector of the columns of \code{df} to show on mouse
hover. Use \code{NULL} to disable hovering. Default: all columns in \code{df} used
for grouping and the \code{feature} to visualize.}

\item{palette}{The palette name or the color function that values will be
mapped to. Default: \code{"inferno"}. Typically one of the following:
\itemize{
\item A character vector of RGB or named colors. Examples: \verb{c("#000000", "#0000FF", "#FFFFFF"))},\verb{topo.colors(10))}.
\item The full name of a RColorBrewer palette, e.g. "BuPu" or "Greens", or
viridis palette: \code{"viridis"}, \code{"magma"}, \code{"inferno"} or \code{"plasma"}.
}

For more options, see parameter \code{palette} of \code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}}.}

\item{zero_values_show}{Logical indicating whether to show deployments with
zero values. Default: \code{TRUE}. See details.}

\item{zero_values_icon_url}{Character with URL to icon for showing
deployments with zero values.  Use \code{NULL} if \code{zero_values_show} is \code{FALSE}.
Default: a cross (multiply symbol)
\code{"https://img.icons8.com/ios-glyphs/30/000000/multiply.png"}. See details.}

\item{zero_values_icon_size}{A number to set the size of the icon to show
groups (e.g. deployments) with zero values. Use \code{NULL} if
\code{zero_values_show} is \code{FALSE}. Default: a cross (multiply symbol)
\code{"https://img.icons8.com/ios-glyphs/30/000000/multiply.png"}. See details.}

\item{na_values_show}{Logical indicating whether to show groups (e.g.
deployments) with no observations when feature is \code{"n_scientificName"}.
Default: \code{TRUE}. See details.}

\item{na_values_icon_url}{Character with URL to icon for showing groups (e.g.
deployments) with no observations. Used only if \code{feature} is
\code{"n_scientificName"}, ignored otherwise. Use \code{NULL} if \code{na_values_show} is
\code{FALSE}. Default: a red cross (multiply symbol)
\code{"https://img.icons8.com/ios-glyphs/30/FA5252/multiply.png"}.}

\item{na_values_icon_size}{A number to set the size of the icon for showing
groups (e.g. deployments) with no observations. Used only if \code{feature} is
\code{"n_scientificName"}, ignored otherwise. Use \code{NULL} if \code{na_values_show} is
\code{FALSE}. Default: 10.}

\item{relative_scale}{Logical indicating whether to use a relative color and
radius scale (\code{TRUE}) or an absolute scale (\code{FALSE}). If absolute scale is
used, specify a valid \code{max_scale}.}

\item{max_scale}{Number indicating the max value used to map color and
radius.}

\item{radius_range}{Vector of length 2 containing the lower and upper limit
of the circle radius. The lower value is used for deployments with zero
feature value, e.g. no observations, no identified species, zero RAI or
zero effort. The upper value is used for the deployment(s) with the highest
feature value (\code{relative_scale} = \code{TRUE}) or \code{max_scale} (\code{relative_scale}
= \code{FALSE}). Default: \code{c(10, 50)}.}
}
\value{
Leaflet map.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=map_summary]{map_summary()}} instead.
}
\details{
Deployments with zero values are shown only if they are present in the
summary, \code{df}. If you want to show deployments with zero values not present
in the summary, be sure to have used \code{extend} = \code{TRUE} in
\code{summarize_observations()}. See the examples.

Notice also that the argument \code{species} is not present anymore. If you want
to visualize a feature only for a specific group of species, please filter
the Camera Trap Data Package using \code{\link[=filter_observations]{filter_observations()}} before using this
function. See the examples.
}
\examples{
x <- example_dataset()
# Filter a data package to get only the data of Anas platyrhynchos
x_anas_p <- x \%>\%
  filter_observations(scientificName == "Anas platyrhynchos")

# Show number of species
x \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude")
  ) \%>\%
  map_summary("n_scientificName")

# Show number of species: show deployments with no
# observations (NA icon and size as defaults)
x_anas_p \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude"),
    extend = TRUE
  ) \%>\%
  map_summary("n_scientificName")

# Show total number of observations (all species)
x \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude")
  ) \%>\%
  map_summary("n_observations")

# Show number of observations of Anas platyrhynchos  - show deployments with
# no observations (zero icon and size  as defaults)
x_anas_p \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude"),
    extend = TRUE
  ) \%>\%
  map_summary("n_observations")

# Show number of observations of Anas platyrhynchos without extending the
# summary: deployments with no observations are not shown
x_anas_p \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude")
  ) \%>\%
  map_summary("n_observations")

# Same result as above if you extend the summary but use
# `zero_values_show` = `FALSE`
x_anas_p \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude"),
    extend = TRUE
  ) \%>\%
  map_summary(
    "n_observations",
    zero_values_show = FALSE,
    zero_values_icon_url = NULL,
    zero_values_icon_size = NULL
  )

# You can eventually filter the species after summarizing,
# although it could not be the most efficient way for large Camera Trap Data
# Packages
library(dplyr)
x \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude", "scientificName")
  ) \%>\%
  filter(scientificName == "Anas platyrhynchos") \%>\%
  map_summary("n_scientificName")

# Show number of individuals
x_anas_p \%>\%
  summarize_observations() \%>\%
  map_summary("sum_count")

# Show RAI
x_anas_p \%>\%
  summarize_observations() \%>\%
  map_summary("rai_observations")

# Show RAI (individual counts)
x_anas_p \%>\%
  summarize_observations() \%>\%
  map_summary("rai_count")

# Show effort (hours)
x_anas_p \%>\%
  summarize_deployments() \%>\%
  map_summary("effort_duration")

# Show effort (days)
x \%>\%
  summarize_deployments() \%>\%
  map_summary("effort_duration", effort_unit = "day")

# Show effort per month
x \%>\%
  summarize_deployments(group_time_by = "month") \%>\%
  map_summary("effort_duration", effort_unit = "day")

# Use viridis palette
x_anas_p \%>\%
  summarize_observations() \%>\%
  map_summary("n_observations", palette = "viridis")

# Use a palette defined by color names
x_anas_p \%>\%
  summarize_observations() \%>\%
  map_summary("n_observations", palette = c("black", "blue", "white"))

# Use a palette defined by hex colors
x_anas_p \%>\%
  summarize_observations() \%>\%
  map_summary("n_observations", palette = c("#000000", "#0000FF", "#FFFFFF"))

# Use same icon but a non default color for zero values deployments,
# E.g. red (hex: E74C3C)
x_anas_p \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude"),
    extend = TRUE
  ) \%>\%
  map_summary(
    "n_observations",
    zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/E74C3C/multiply.png"
  )

# Use another icon via a different URL, e.g. the character Fry from Futurama
# in green (2ECC71)
x_anas_p \%>\%
  summarize_observations(extend = TRUE) \%>\%
  map_summary(
    "n_observations",
    zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/2ECC71/futurama-fry.png"
  )

# Set size of the icon for zero values deployments
x \%>\%
  filter_observations(scientificName == "Anas platyrhynchos") \%>\%
  summarize_observations(extend = TRUE) \%>\%
  map_summary("n_observations", zero_values_icon_size = 30)

# Use another icon url/size for visualizing groups with no observations
# (NA, only for `n_scientificName` feature)
x \%>\%
  filter_observations(deploymentID != "00a2c20d") \%>\%
  summarize_observations(
  group_by = c("deploymentID", "latitude", "longitude")
  ) \%>\%
  map_summary(
    feature = "n_scientificName",
    na_values_icon_url = "https://img.icons8.com/ios-glyphs/30/E74C3C/futurama-fry.png",
    na_values_icon_size = 60
  )

# Disable cluster
x \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude")
  ) \%>\%
  map_summary("n_observations", cluster = FALSE)

# Show only number of observations and location name while hovering
x \%>\%
  summarize_observations(
    group_by = c("deploymentID", "locationName", "latitude", "longitude")
  ) \%>\%
  map_summary(
    "n_observations",
    hover_columns = c("locationName", "n_observations")
)

# Use absolute scale for colors and radius
x \%>\%
  summarize_observations(
    group_by = c("deploymentID", "latitude", "longitude")
  ) \%>\%
  map_summary(
    "n_scientificName",
    relative_scale = FALSE,
    max_scale = 4
)

# Change max and min size circles
x \%>\%
  summarize_observations(
    group_by = c("deploymentID", "locationName", "latitude", "longitude")
  ) \%>\%
  map_summary(
    "n_observations",
    radius_range = c(40, 150)
)
}
\concept{deprecated visualization functions}
