% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R
\name{mica}
\alias{mica}
\title{Deprecated datasets in camtraptor}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These datasets are deprecated:
\itemize{
\item Use \code{\link[=example_dataset]{example_dataset()}} instead of \code{mica}.
\item Use \code{animal_positions} instead of \code{animal_pos}. Note that the column
\code{sequenceID} has been renamed to \code{eventID}.
\item Use \code{calibration_models} instead of \code{calib_models}.
}
}
\seealso{
Other sample data:
\code{\link{animal_pos}},
\code{\link{animal_positions}},
\code{\link{calib_models}},
\code{\link{calibration_models}}

Other read sample data functions:
\code{\link[=example_dataset]{example_dataset()}}
}
\concept{read sample data functions}
\concept{sample data}
