% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pos.R
\name{predict_pos}
\alias{predict_pos}
\title{Predict position}
\usage{
predict_pos(
  animal_pos,
  calib_models,
  dep_tag = "deployment",
  x = "x",
  y = "y",
  image_width = "ImageWidth",
  image_height = "ImageHeight"
)
}
\arguments{
\item{animal_pos}{Dataframe of animal position digitisation data.
It must contain (at least) the columns defined in args \code{x}, \code{y}, \code{dep_tag},
\code{image_width}, \code{image_height}.}

\item{calib_models}{Named list of deployment calibration models or site calibration
models (\code{calibs} objects), produced using \code{cal.site()} (not yet included in
this package). The deployment names are used as names.}

\item{dep_tag}{Character naming the column within \code{animal_pos} against which
names of the elements can be matched to apply the right deployment
calibration models. Default: \code{"deployment"}.}

\item{x}{Character naming the column within \code{animal_pos} containing x pixel
positions for each digitised point. Default: \code{"x"}.}

\item{y}{Character naming the column within \code{animal_pos} containing y pixel
positions for each digitised point. Default: \code{"y"}.}

\item{image_width}{pixel x dimension of each image.
It must be consistent for each deployment.}

\item{image_height}{pixel y dimension of each image.
It must be consistent for each deployment.}
}
\value{
Original (tibble) dataframe as passed via \code{animal_pos} with
additional columns:
\itemize{
\item \code{radius}: radial distance from camera
\item \code{angle}: angular distance from camera
\item \code{frame_count}: indicator of the frame order within each sequence
}
}
\description{
Predicts position relative to camera given image pixel positions and site
calibration models.
}
\examples{
predict_pos(animal_positions, dep_calib_models)
}
