% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{read_camtrapdp}
\alias{read_camtrapdp}
\title{Read a Camera Trap Data Package}
\usage{
read_camtrapdp(file)
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}
}
\description{
Re-export of \code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}. See that function for full
documentation.
}
\concept{read functions}
