% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_coordinates.R
\name{round_coordinates}
\alias{round_coordinates}
\title{Round coordinates to generalize camera trap locations}
\usage{
round_coordinates(x, digits = 3)
}
\arguments{
\item{x}{A Camtrap DP, as read by \code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{digits}{Number of decimal places to round coordinates to (\code{1}, \code{2} or
\code{3}).}
}
\value{
\code{x} with rounded coordinates as well as updated
\code{coordinateUncertainty}.(in deployments) and \code{coordinatePrecision} (in
metadata).
}
\description{
Rounds deployment coordinates to a certain number of digits to
fuzzy/generalize camera trap locations.
This function can be used before publishing data in order to protect
sensitive species and/or prevent theft of active cameras.
}
\section{Details}{

Rounding coordinates is a recommended method to generalize sensitive
biodiversity information (see
\href{https://doi.org/10.15468/doc-5jp4-5g10#s-generalization}{Section 4.2}
in Chapman 2020).
Choose a number of digits that aligns with the sensitivity of the data and
notice the effect on precision and uncertainty.
Publish the coordinates as is (i.e. do not use this function) if the data
are not sensitive.\tabular{llll}{
   sensitivity \tab digits \tab coordinatePrecision \tab coordinateUncertainty \cr
   high \tab 1 \tab 0.1 \tab original uncertainty + 15691 m \cr
   medium \tab 2 \tab 0.01 \tab original uncertainty + 1570 m \cr
   low \tab 3 \tab 0.001 \tab original uncertainty + 157 m \cr
}


For records with \code{coordinateUncertainty = NA} the function will assume the
coordinates were obtained by GPS and use \verb{30 m} as original uncertainty,
before adding uncertainty caused by rounding.
The added uncertainty is the largest possible value caused by rounding (see
\href{https://doi.org/10.15468/doc-gg7h-s853#table-uncertainty}{Table 3} in
Chapman & Wieczorek 2020).
}

\examples{
# Round coordinates of example package to 3 digits
mica <- round_coordinates(mica, 3)

# coordinatePrecision is set in metadata
mica$coordinatePrecision

# coordinateUncertainty is set in data: original uncertainty (or 30) + 157 m
deployments(mica)$coordinateUncertainty
}
\concept{publication functions}
