% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_deployments.R
\name{summarize_deployments}
\alias{summarize_deployments}
\alias{summarise_deployments}
\title{Summarize deployments information}
\usage{
summarize_deployments(
  x,
  group_by = c("deploymentID", "latitude", "longitude"),
  group_time_by = NULL
)

summarise_deployments(
  x,
  group_by = c("deploymentID", "latitude", "longitude"),
  group_time_by = NULL
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{group_by}{Character vector with the names of the columns in
deployments. At the moment you can choose one or many columns among:
\code{c("deploymentID", "latitude", "longitude", "locationID", "locationName", "deploymentStart", "deploymentEnd", "deploymentTags")}. Default:
\code{c("deploymentID", "latitude", "longitude")}.}

\item{group_time_by}{Character, one of \code{"day"}, \code{"week"}, \code{"month"},
\code{"year"}. The effort is calculated at the interval rate defined in
\code{group_time_by}. Default: \code{NULL}, no grouping, i.e. the entire duration of
the deployment is taken into account as a whole.}
}
\value{
A grouped tibble data frame with the following columns:
\itemize{
\item \code{group_by} names, e.g. \code{deploymentID}, \code{latitude}, \code{longitude} and
\code{locationName}.
\item \code{group_time_by} name if provided, e.g. \code{month}. It contains the first
date of the time interval, e.g. the first day of the month.
\item \code{effort_duration}: A duration object (duration is a class from lubridate
package). Duration is always recorded as a fixed number of seconds. See
\code{\link[lubridate:duration]{lubridate::duration()}}.
}
}
\description{
Summarizes deployments information, more specifically the duration effort.
}
\details{
\code{summarize_deployments()} and \code{summarise_deployments()} are synonyms.
}
\examples{
x <- example_dataset()

# Return effort using default `group_by` and no time grouping
summarize_deployments(x)

# Return effort using default `group_by` and grouping by year
summarize_deployments(x, group_time_by = "year")

# Return effort specifying grouping columns, e.g. `deploymentID` and
# `locationName` and grouping by day
summarize_deployments(
  x,
  group_by = c("deploymentID", "locationName"),
  group_time_by = "day"
)

# Afterwards, you can calculate the total effort over all deployments. You 
# can also show other information, e.g. the (number of) deployments and
# locations.
library(dplyr)
summarize_deployments(
  x,
  group_by = c("deploymentID", "locationName"),
  group_time_by = "month"
) \%>\%
  group_by(month) \%>\%
  summarise(
    deploymentIDs = list(deploymentID),
    ndep = length(unique(deploymentID)),
    nloc = length(unique(locationName)),
    effort_duration = sum(effort_duration)
  )
}
\seealso{
Other exploration functions:
\code{\link[=summarize_observations]{summarize_observations()}}
}
\concept{exploration functions}
