% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{transform_effort_to_common_units}
\alias{transform_effort_to_common_units}
\title{Transform efforts to common units.}
\usage{
transform_effort_to_common_units(effort, unit)
}
\arguments{
\item{effort}{a vector of duration objects}

\item{unit}{common unit to express duration objects. One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
}}
}
\value{
a numeric vector
}
\description{
This function is useful for visualization and communication purposes. Efforts
are duration objects and so they are always expressed in seconds. Although
they are also returned on screen in common units, e.g. "2594308s (~4.29
weeks)", the values in seconds are used in color scales, not very handy to be
interpreted. Converting them in the most suitable time unit is also useful
for communication purposes (reports, research articles, ...). Obviously the
conversion can be not always precise, e.g. month is not always 30 days long.
}
\examples{
# create efforts (durations) to transform
efforts <- c(lubridate::duration("2hours 2minutes 1second"),
            lubridate::duration("3days 2hours"))
# transform effort to hours
get_effort_in_common_units(efforts, "hour")
# transform effort to days
get_effort_in_common_units(efforts, "day")
}
