% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_species_per_deployment.R
\name{visualize_species_per_deployment}
\alias{visualize_species_per_deployment}
\title{Visualize number of species per deployment}
\usage{
visualize_species_per_deployment(deployments, observations, cluster = TRUE)
}
\arguments{
\item{deployments}{a tibble (data.frame) containing deployments}

\item{observations}{a tibble (data.frame) containing observations}

\item{cluster}{a logical value indicating whether using the cluster option
while visualizing maps. Default: TRUE}
}
\value{
a leaflet map
}
\description{
This function visualizes the number of species detected by each deployment.
Possible filtering about time period, ... (to be added)
}
