% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc_wi.R
\name{write_dwc_wi}
\alias{write_dwc_wi}
\title{Transform Wildlife Insight camera trap data to Darwin Core}
\usage{
write_dwc_wi(
  import_directory = ".",
  export_directory = ".",
  rights_holder = "project_admin_organization",
  coordinateUncertaintyInMeters = 30
)
}
\arguments{
\item{import_directory}{Path to local directory to read the WI files}

\item{export_directory}{Path to local directory to write files to.}

\item{rights_holder}{Acronym of the organization owning or managing the
rights over the data.}

\item{coordinateUncertaintyInMeters}{Uncertainty of the coordinate in meters.}
}
\value{
CSV (data) files written to disk.
}
\description{
Transforms a \href{https://www.wildlifeinsights.org/}{Wildlife Insight (WI)}
dataset to Darwin Core CSV that can be uploaded to a
\href{https://www.gbif.org/ipt}{GBIF IPT} for publication.
}
\details{
Wildlife Insight offers to export your dataset \href{https://www.wildlifeinsights.org/get-started/download/private}{as private download} or
\href{https://www.wildlifeinsights.org/get-started/data-download/public}{public download}.
This function read the zip file from the download and convert the Projects,
Cameras, Deployements csv file to Darwin Core standard
}
