% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eml.R
\name{write_eml}
\alias{write_eml}
\title{Transform Camtrap DP metadata to EML}
\usage{
write_eml(
  package,
  directory = ".",
  title = package$title,
  description = package$description,
  creators = NULL,
  keywords = c("camera traps")
)
}
\arguments{
\item{package}{A Camtrap DP, as read by \code{\link[=read_camtrap_dp]{read_camtrap_dp()}}.}

\item{directory}{Path to local directory to write file to.
If \code{NULL}, then the EML object is returned instead, which can be useful
for extended/adapting the EML before writing with \code{\link[EML:write_eml]{EML::write_eml()}}.}

\item{title}{Dataset title.}

\item{description}{Dataset description.
Will be added after an automatically generated paragraph.
Multiple paragraphs can be provided as a character vector.}

\item{creators}{Dataset creators
\itemize{
\item If \code{NULL} then all \code{package$contributors} will be added as creators, in the
order as listed.
\item If e.g. \code{c("Emma Cartuyvels", "Jim Casaer", "...", "Peter Desmet")}, then
Emma Cartuyvels, Jim Casaer and Peter Desmet will be set as first, second
and last creators respectively, on the condition that their name (\code{title})
is present in \code{package$contributors}.
All other contributors will be inserted at \code{"..."}, sorted on their last
name.
}}

\item{keywords}{Dataset keywords.}
}
\value{
\code{eml.xml} file written to disk or \code{EML} object when
\code{directory = NULL}.
}
\description{
Transforms the metadata of a \href{https://tdwg.github.io/camtrap-dp/}{Camera Trap Data Package (Camtrap DP)} to an
\href{https://eml.ecoinformatics.org/}{EML} file that can be uploaded to a
\href{https://www.gbif.org/ipt}{GBIF IPT} for publication.
}
\section{Transformation details}{

Metadata is derived from what is provided in \code{package} and in the function
parameters.
The following properties are set:
\itemize{
\item \strong{title}: Title as provided in \code{title} or \code{package$title}.
\item \strong{description}: Description as provided in \code{description} or
\code{package$description}.
The description is preceded by an automatically generated paragraph
describing from which project and platform the dataset is derived.
\item \strong{license}: License with scope \code{data} as provided in \code{package$licenses}.
\item \strong{creators}: Contributors (all roles) as provided in
\code{package$contributors}, filtered/reordered based on \code{creators}.
\item \strong{contact}: First creator.
\item \strong{metadata provider}: First creator.
\item \strong{keywords}: Keywords as provided in \code{keywords}.
\item \strong{associated parties}: Organizations as provided in
\code{package$organizations}.
\item \strong{geographic coverage}: Bounding box as provided in \code{package$spatial}.
\item \strong{taxonomic coverage}: Species (no other ranks) as provided in
\code{package$taxonomic}.
\item \strong{temporal coverage}: Date range as provided in \code{package$temporal}.
\item \strong{project data}: Title, acronym as identifier, description, and sampling
design as provided in \code{package$project}.
The first creator is set as project personnel.
\item \strong{alternative identifier}: Identifier as provided in \code{package$id}.
If this is a DOI, no new DOI will be created when publishing to GBIF.
\item \strong{external link}: URL of the project as provided in
\code{package$project$path}.
}

To be set manually in the GBIF IPT: \strong{type}, \strong{subtype},
\strong{update frequency} and \strong{publishing organization}.

Not set: \strong{sampling methods} and \strong{citations}.

Not applicable: \strong{collection data}.
}

