% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinates.R
\name{add_coordinates}
\alias{add_coordinates}
\title{Add deployment coordinates to observations}
\usage{
add_coordinates(x)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
Camera trap data package object, where \code{observations} is updated by
appending two new columns: \code{latitude} and \code{longitude}
}
\description{
This function adds deployment coordinates to \code{observations} based on
\code{deploymentID}.
}
\examples{
x <- example_dataset()

# Add coordinates to observations
add_coordinates(x) \%>\% observations()
}
\seealso{
Other accessor functions: 
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()},
\code{\link{taxa}()}
}
\concept{accessor functions}
