% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_positions}
\alias{animal_positions}
\title{Sample of animal position digitization data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 6 columns.
}
\usage{
animal_positions
}
\description{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}
\item \code{eventID}
\item \code{x} and \code{y}: The coordinates.
\item \code{imageWidth} and \code{imageHeight}: The image dimensions.
}
}
\seealso{
Other sample data:
\code{\link{calibration_models}},
\code{\link{camtraptor-deprecated}}
}
\concept{sample data}
\keyword{datasets}
