% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_positions}
\alias{animal_positions}
\title{Sample of animal position digitisation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 6 columns.
}
\usage{
animal_positions
}
\description{
It must contain at least columns containing the coordinates (x, y), the
deployment and the image dimensions (width and height).
}
\keyword{datasets}
