% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_animal_pos.R
\name{calc_animal_pos}
\alias{calc_animal_pos}
\title{Calculate animal position}
\usage{
calc_animal_pos(
  animal_pos,
  calib_models,
  dep_tag = "deploymentID",
  sequence_id = "sequenceID",
  x = "x",
  y = "y",
  image_width = "imageWidth",
  image_height = "imageHeight"
)
}
\arguments{
\item{animal_pos}{Data.frame (tibble) of animal position digitisation data.
It must contain (at least) the columns defined in args \code{dep_tag},
\code{sequence_id}, \code{x}, \code{y}, \code{image_width} and \code{image_height}.}

\item{calib_models}{Named list of deployment calibration models or site calibration
models (\code{calibs} objects), produced using \code{cal.site()} (not yet included in
this package). The deployment names are used as names.}

\item{dep_tag}{Column in \code{animal_pos} against which
names of the elements can be matched to apply the right deployment
calibration models. Default: \code{"deploymentID"}.}

\item{sequence_id}{Column in \code{animal_pos} containing
the sequence ID the images belong to. Default: \code{"sequenceID"}.}

\item{x}{Column in \code{animal_pos} containing x pixel
positions for each digitised point. Default: \code{"x"}.}

\item{y}{Column in \code{animal_pos} containing y pixel
positions for each digitised point. Default: \code{"y"}.}

\item{image_width}{Column in \code{animal_pos} containing the pixel x dimension of
each image. Default: \code{"imageWidth"}. Notice that the pixel x dimension must
be consistent for each deployment.}

\item{image_height}{Column in \code{animal_pos} containing the pixel y dimension
of each image. Default: \code{"imageHeight"}. Notice that the pixel y dimension
must be consistent for each deployment.}
}
\value{
Original (tibble) data.frame as passed via \code{animal_pos} with
additional columns:
\itemize{
\item \code{radius}: radial distance from camera
\item \code{angle}: angular distance from camera
\item \code{frame_count}: indicator of the frame order within each sequence
}
}
\description{
Calculate position relative to camera given image pixel positions and site
calibration models.
}
\examples{
# use default values
calc_animal_pos(animal_positions, dep_calib_models)
}
