% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_animal_pos.R
\name{calc_animal_pos}
\alias{calc_animal_pos}
\title{This function is deprecated. Please use \code{\link[=calculate_individual_radius_angle]{calculate_individual_radius_angle()}}
instead.}
\usage{
calc_animal_pos(
  animal_pos,
  calib_models,
  dep_tag = "deploymentID",
  sequence_id = "sequenceID",
  x = "x",
  y = "y",
  image_width = "imageWidth",
  image_height = "imageHeight"
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
A matrix. Row names always indicate the station ID. Column names are
dates.
}
\description{
Returns the \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camera operation matrix} as
returned by \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camtrapR::cameraOperation()}.
}
\details{
The deployment data are by default grouped by \code{locationName} (station ID in
camtrapR jargon) or another column specified by the user via the
\code{station_col} argument. If multiple deployments are linked to same location,
daily efforts higher than 1 occur.

Partially active days, e.g. the first or the last day of a deployment, result
in decimal effort values as in \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camtrapR::cameraOperation()}.
}
\examples{
x <- example_dataset()
camtrapR_cameraOperation(x)

# Specify column with station names
camtrapR_cameraOperation(x, station_col = "locationID")

# Specify column with session IDs
x_sessions <- x
x_sessions$data$deployments <- deployments(x_sessions) \%>\%
  dplyr::mutate(session = ifelse(
    stringr::str_starts(.data$locationName, "B_DL_"),
      "after2020",
      "before2020"
  )
)
camtrapR_cameraOperation(x_sessions, session_col = "session")

# Specify column with camera IDs
x_cameras <- x_sessions
x_cameras$data$deployments$cameraID <- c(1, 2, 3, 4)
camtrapR_cameraOperation(x_cameras, camera_col = "cameraID")

# Specify both session and camera IDs
camtrapR_cameraOperation(
  x_cameras,
  camera_col = "cameraID",
  session_col = "session"
)

# Use prefix Station as in camtrapR's camera operation matrix
camtrapR_cameraOperation(x, use_prefix = TRUE)
}
\seealso{
Other camtrapR-derived functions: 
\code{\link{camtrapR_detectionHistory}()},
\code{\link{camtrapR_recordTable}()}
}
