% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R
\name{calib_models}
\alias{calib_models}
\title{Deprecated datasets in camtraptor}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item The \code{mica} dataset was a Camtrap DP 0.1.6 data package and it is not
supported anymore. Replaced by \code{\link[=example_dataset]{example_dataset()}}.
\item The sample of animal position digitization data \code{animal_pos} is deprecated.
Use \code{\link{animal_positions}} instead. The column \code{"sequenceID"} has been renamed to \code{"eventID"}.
\item The sample of deployment calibration models \code{calib_models} is deprecated.
Use \code{\link{calibration_models}} instead.
}
}
\seealso{
Other sample data:
\code{\link{animal_pos}},
\code{\link{animal_positions}},
\code{\link{calibration_models}},
\code{\link{mica}}
}
\concept{sample data}
