% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration_models}
\alias{calibration_models}
\title{Sample of deployment calibration models}
\format{
An object of class \code{calibs} of length 4.
}
\usage{
calibration_models
}
\description{
A list containing a number of calibration models (\code{calibs}) or site
calibration models (\code{depcal}).
The deployment names are used as names.
}
\seealso{
Other sample data:
\code{\link{animal_pos}},
\code{\link{animal_positions}},
\code{\link{calib_models}},
\code{\link{mica}}
}
\concept{sample data}
\keyword{datasets}
