% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtraptor-package.R
\docType{package}
\name{camtraptor-package}
\alias{camtraptor}
\alias{camtraptor-package}
\title{camtraptor: Explore and Visualize Camera Trap Data Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Explore and visualize Camera Trap Data Packages ('Camtrap DP', \url{https://camtrap-dp.tdwg.org/}). camtraptor offers a step-by-step workflow to read Camtrap DP files, filter data of interest, summarize information (e.g. number of observed species) and visualize this per deployment on an interactive map. You can also use it to transform data for analysis in 'camtrapR'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/camtraptor}
  \item \url{https://inbo.github.io/camtraptor/}
  \item Report bugs at \url{https://github.com/inbo/camtraptor/issues}
}

}
\author{
\strong{Maintainer}: Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))

Authors:
\itemize{
  \item Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
  \item Sanne Govaert \email{sanne.govaert@inbo.be} (\href{https://orcid.org/0000-0002-8939-1305}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
  \item Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
  \item Wolf Missotten \email{wolfmissotten@gmail.com} (\href{https://orcid.org/0009-0006-4069-0685}{ORCID}) (affiliation: Ghent University)
}

Other contributors:
\itemize{
  \item Martijn Bollen \email{martijn.bollen@inbo.be} (\href{https://orcid.org/0000-0002-1621-6258}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]
  \item Jim Casaer \email{jim.casaer@inbo.be} (\href{https://orcid.org/0000-0001-6788-5876}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]
  \item Nele Mullens (\href{https://orcid.org/0000-0002-3054-9665}{ORCID}) (affiliation: University of Antwerp) [contributor]
  \item Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-1020}{ORCID}) (affilication: Swiss Ornithological Institute) [contributor]
  \item Thierry Onkelinx \email{thierry.onkelinx@inbo.be} (\href{https://orcid.org/0000-0001-8804-4216}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]
  \item Lynn Pallemaerts \email{lynn.pallemaerts@inbo.be} (\href{https://orcid.org/0000-0002-5034-5416}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]
  \item Research Institute for Nature and Forest (INBO) (\href{https://ror.org/00j54wy13}{ROR}) [copyright holder]
  \item Research Foundation - Flanders (https://lifewatch.be) [funder]
}

}
\keyword{internal}
