% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtraptor-package.R
\docType{package}
\name{camtraptor-package}
\alias{camtraptor}
\alias{camtraptor-package}
\title{camtraptor: Read, Explore and Visualize Camera Trap Data Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Read, explore and visualize Camera Trap Data Packages (Camtrap DP). 'Camtrap DP' (\url{https://camtrap-dp.tdwg.org}) is a community developed data exchange format for this type of data. With camtraptor you can read and filter data, create overviews of observed species, relative abundance or effort, and plot these data on a map.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/inbo/camtraptor/issues}
}

}
\author{
\strong{Maintainer}: Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID})

Authors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID})
  \item Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID})
}

Other contributors:
\itemize{
  \item Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-10205}{ORCID}) [contributor]
  \item Thierry Onkelinx \email{thierry.onkelinx@inbo.be} (\href{https://orcid.org/0000-0001-8804-4216}{ORCID}) [contributor]
  \item Jim Casaer \email{jim.casaer@inbo.be} (\href{https://orcid.org/0000-0001-6788-5876}{ORCID}) [contributor]
  \item Lynn Pallemaerts \email{lynn.pallemaerts@inbo.be} (\href{https://orcid.org/0000-0002-5034-5416}{ORCID}) [contributor]
  \item Nele Mullens (\href{https://orcid.org/0000-0002-3054-9665}{ORCID}) [contributor]
  \item Research Institute for Nature and Forest (INBO) (https://www.vlaanderen.be/inbo/en-gb/) [copyright holder]
  \item LifeWatch Belgium (https://lifewatch.be) [funder]
}

}
\keyword{internal}
