% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{check_filter_symbol}
\alias{check_filter_symbol}
\title{Check filter symbol}
\usage{
check_filter_symbol(symbol)
}
\arguments{
\item{symbol}{character with symbol for filter predicate, e.g. "=="}
}
\description{
Check filter symbol
}
