% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{check_filter_type}
\alias{check_filter_type}
\title{Check filter type}
\usage{
check_filter_type(type)
}
\arguments{
\item{type}{character with type for filter predicate, e.g. "equals"}
}
\description{
Check filter type
}
