% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species.R
\name{check_species}
\alias{check_species}
\title{Check scientific or vernacular name(s)}
\usage{
check_species(
  package = NULL,
  species,
  arg_name = "species",
  datapkg = lifecycle::deprecated()
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{species}{Character vector with scientific or vernacular names.}

\item{arg_name}{Character with argument name to return in error message
Default: "species".}

\item{datapkg}{Deprecated. Use \code{package} instead.}
}
\value{
A character vector with the correspondent scientific names.
}
\description{
Checks if a given scientific or vernacular name(s) can be found in the
metadata (\code{package$taxonomic}) and returns error if not.
}
\examples{
# Species is a scientific name
check_species(mica, "Martes foina")

# Species is a vector of vernacular names
check_species(mica, c("beech marten", "european polecat"))

# Vernacular names can be specified in any language available
check_species(mica, c("vos", "blauwe reiger"))

# Vernacular names and scientific names can be mixed up
check_species(mica, c("beech marten", "blauwe reiger", "Anas strepera"))

# Case insensitive
check_species(mica, "AnaS StrePeRa")
check_species(mica, "bEEch mARteN")

\dontrun{
check_species(mica, "bad name")
}
}
\concept{validation functions}
