% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species.R
\name{check_species}
\alias{check_species}
\title{Check scientific or vernacular name(s)}
\usage{
check_species(x, species, arg_name = "species")
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{species}{Character vector with scientific or vernacular names.}

\item{arg_name}{Character with argument name to return in error message
Default: "species".}
}
\value{
A character vector with the correspondent scientific names.
}
\description{
Checks if a given scientific or vernacular name(s) can be found in the
metadata (\code{x$taxonomic}) and returns error if not.
}
\examples{
# Species is a scientific name
check_species(mica, "Martes foina")

# Species is a vector of vernacular names
check_species(mica, c("beech marten", "european polecat"))

# Vernacular names can be specified in any language available
check_species(mica, c("vos", "blauwe reiger"))

# Vernacular names and scientific names can be mixed up
check_species(mica, c("beech marten", "blauwe reiger", "Anas strepera"))

# Case insensitive
check_species(mica, "AnaS StrePeRa")
check_species(mica, "bEEch mARteN")

\dontrun{
check_species(mica, "bad name")
}
}
\concept{validation functions}
