% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species.R
\name{check_species}
\alias{check_species}
\title{Check validity of the given scientific or vernacular name(s)}
\usage{
check_species(datapkg, species, arg_name = "species")
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations} 2. \code{deployments} 3. \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{species}{a character vector with scientific or vernacular names.}

\item{arg_name}{character with argument name to return in error message.
Default: "species".}
}
\value{
a character vector with the correspondent scientific names
}
\description{
Check names against all scientific and vernacular names contained in a camera
trap data package
}
\examples{
check_species(camtrapdp, "Gallinula chloropus")
check_species(camtrapdp, c("Mallard", "Norway Rat"))
# case insensitive
check_species(camtrapdp, "galliNULa CHloropUs")
check_species(camtrapdp, "MalLARD")

}
