% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{pred}
\alias{pred_not}
\alias{pred_gt}
\alias{pred_gte}
\alias{pred_lt}
\alias{pred_lte}
\alias{pred_in}
\alias{pred_notin}
\alias{pred_na}
\alias{pred_notna}
\alias{pred_and}
\alias{pred_or}
\alias{apply_filter_predicate}
\alias{check_species}
\alias{get_scientific_name}
\alias{read_wi}
\title{Defunct functions}
\usage{
# Deprecated in 1.0.0 -------------------------------------

pred(arg, value)

pred_not(arg, value)

pred_gt(arg, value)

pred_gte(arg, value)

pred_lt(arg, value)

pred_lte(arg, value)

pred_in(arg, value)

pred_notin(arg, value)

pred_na(arg, value)

pred_notna(arg, value)

pred_and(arg, value)

pred_or(arg, value)

apply_filter_predicate(arg, value)

check_species(x)

get_scientific_name(x)

read_wi(x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions are not supported anymore. If there's a known replacement,
calling the function will tell you about it.
}
\keyword{internal}
