% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployments.R
\name{deployments}
\alias{deployments}
\title{Retrieve deployment data from a Camera Trap Data Package}
\usage{
deployments(package)
}
\arguments{
\item{package}{}
}
\value{
A tibble with the deployments of the \code{package}
}
\description{
Retrieve deployment data from a Camera Trap Data Package
}
\examples{
deployments(mica)
}
