% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{.map_summary_legend_titles}
\alias{.map_summary_legend_titles}
\title{Map legend title table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\usage{
.map_summary_legend_titles
}
\description{
Store legend titles for deployment visualizations: RAI, effort, number of
observations, etc.
A data frame of all titles with the following columns:
\itemize{
\item \code{feature}: Deployment feature to visualize.
\item \code{legend_title}: Legend title.
}
}
\keyword{internal}
