% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{events}
\alias{events}
\title{Get events}
\usage{
events(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the events, containing the
following columns:
\itemize{
\item \code{deploymentID}
\item \code{eventID}
\item \code{eventStart}
\item \code{eventEnd}
}
}
\description{
Re-export of \code{\link[camtrapdp:events]{camtrapdp::events()}}.
}
\examples{
x <- example_dataset()
events(x)
}
\seealso{
Other accessor functions:
\code{\link[=contributors]{contributors()}},
\code{\link[=deployments]{deployments()}},
\code{\link[=individuals]{individuals()}},
\code{\link[=locations]{locations()}},
\code{\link[=media]{media()}},
\code{\link[=observations]{observations()}},
\code{\link[=taxa]{taxa()}}
}
\concept{accessor functions}
