% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{filter_deployments}
\alias{filter_deployments}
\title{Filter deployments}
\usage{
filter_deployments(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}

\item{...}{Filtering conditions, see \code{dplyr::filter()}.}
}
\description{
Re-export of \code{\link[camtrapdp:filter_deployments]{camtrapdp::filter_deployments()}}. See that function for full
documentation.
}
\seealso{
Other filter functions: 
\code{\link{filter_media}()},
\code{\link{filter_observations}()},
\code{\link{filter_out_timelapse}()}
}
\concept{filter functions}
