% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{filter_media}
\alias{filter_media}
\title{Filter media}
\usage{
filter_media(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}

\item{...}{Filtering conditions, see \code{dplyr::filter()}.}
}
\description{
Re-export of \code{\link[camtrapdp:filter_media]{camtrapdp::filter_media()}}. See that function for full
documentation.
}
\seealso{
Other filter functions: 
\code{\link{filter_deployments}()},
\code{\link{filter_observations}()},
\code{\link{filter_out_timelapse}()}
}
\concept{filter functions}
