% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{filter_observations}
\alias{filter_observations}
\title{Filter observations}
\usage{
filter_observations(x, ...)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}

\item{...}{Filtering conditions, see \code{dplyr::filter()}.}
}
\description{
Re-export of \code{\link[camtrapdp:filter_observations]{camtrapdp::filter_observations()}}. See that function for full
documentation.
}
\seealso{
Other filter functions: 
\code{\link{filter_deployments}()},
\code{\link{filter_media}()},
\code{\link{filter_out_timelapse}()}
}
\concept{filter functions}
