% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cam_op.R
\name{get_cam_op}
\alias{get_cam_op}
\title{Get camera operation matrix}
\usage{
get_cam_op(datapkg, ...)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{...}{filter predicates for filtering on deployments}
}
\value{
a matrix. Row names always indicate the \code{location_name} (station ID)
\code{"Station"}+\code{location_name}. Column names are dates.
}
\description{
This function returns the \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camera operation matrix}
as returned by camtrapR's function \code{cameraOperation()}.
}
\details{
The deployment data are grouped by \code{location_name} (station ID in camtrapR
jargon). If multiple deploymnts are linked to same location, daily
efforts higher than one occur.
}
\examples{
get_cam_op(camtrapdp)
# applying filter(s) on deployments, e.g. deployments with latitude >= 51.28
get_cam_op(camtrapdp, pred_gte("latitude", 51.28))
}
