% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cam_op.R
\name{get_cam_op}
\alias{get_cam_op}
\title{Get camera operation matrix}
\usage{
get_cam_op(datapkg, ..., station_col = "locationName", use_prefix = FALSE)
}
\arguments{
\item{datapkg}{A camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates for filtering on deployments.}

\item{station_col}{Column name to use for identifying the stations. Default:
\code{"locationName"}.}

\item{use_prefix}{logical (\code{TRUE}or \code{FALSE}). If \code{TRUE} the returned row
names will start with prefix \code{"Station"} as returned by
\code{camtrapR::cameraOperation()}. Default: \code{FALSE}.}
}
\value{
a matrix. Row names always indicate the station ID. Column names are
dates.
}
\description{
This function returns the \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camera operation matrix}
as returned by camtrapR's function \code{cameraOperation()}.
}
\details{
The deployment data are by default grouped by \code{locationName} (station ID in
camtrapR jargon) or another column specified by the user. If multiple
deployments are linked to same location, daily efforts higher than 1 occur.

Partially active days, e.g. the first or the last day of a deployment result
in decimal effort values, same behavior as camtrapR's function \code{cameraOperation()}.
}
\examples{
get_cam_op(mica)
# Applying filter(s) on deployments, e.g. deployments with latitude >= 51.18
get_cam_op(mica, pred_gte("latitude", 51.18))
# Specify column with station names
get_cam_op(mica, station_col = "locationID")
# Use prefix Station as in camtrapR's camera operation matrix
get_cam_op(mica, use_prefix = TRUE)
}
\seealso{
Other get_functions: 
\code{\link{filter_predicate}},
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_record_table}()}
}
\concept{get_functions}
