% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cam_op.R
\name{get_cam_op}
\alias{get_cam_op}
\title{Get camera operation matrix}
\usage{
get_cam_op(datapkg)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations} 2. \code{deployments} 3. \code{multimedia}
}

and a list with metadata: \code{datapackage}}
}
\value{
a matrix. Row names always indicate the \code{location_name} (station) and
the \code{camera_id} in the form
\code{"Station"}+\code{location_name}\link[=+`"__CAM_"`+`camera_id`]{+\code{"__CAM_"}+\code{camera_id}} if \code{camera_id} is specified in \code{deployments}. Column names are
dates.
}
\description{
This function returns the \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camera operation matrix}
as returned by \code{camtrapR::cameraOperation()}.
}
