% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cam_op.R
\name{get_cam_op}
\alias{get_cam_op}
\title{Get camera operation matrix}
\usage{
get_cam_op(datapkg)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations} 2. \code{deployments} 3. \code{multimedia}
}

and a list with metadata: \code{datapackage}}
}
\value{
a matrix. Row names always indicate the \code{location_name} (station)
\code{"Station"}+\code{location_name}. Column names are dates.
}
\description{
This function returns the \href{https://jniedballa.github.io/camtrapR/reference/cameraOperation.html}{camera operation matrix}
as returned by \code{camtrapR::cameraOperation()}.
}
\details{
The deployment data are grouped by \code{location_name} (station in camtrapR
jargon) and the function assumes there was 1 camera per location (station).
}
