% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_effort.R
\name{get_custom_effort}
\alias{get_custom_effort}
\title{Get custom effort}
\usage{
get_custom_effort(
  x,
  ...,
  start = NULL,
  end = NULL,
  group_by = NULL,
  unit = "hour"
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{...}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} filter predicates for filtering
on deployments are not supported anymore and an error is returned. Anything
else throws a deprecation warning and will be ignored. Please use
\code{filter_deployments()} to filter on deployments.}

\item{start}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} if not \code{NULL }. Not supported
anymore. Use \code{filter_deployments()} to filter out deployments.}

\item{end}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} if not \code{NULL }. Not supported
anymore. Use \code{filter_deployments()} to filter out deployments.}

\item{group_by}{Character, one of \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"}.
Default: \code{NULL}. See \code{group_time_by} argument in
\verb{[summarize_deployments()]}.}

\item{unit}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The unit used to quantify the
effort. Ignored as the effort is returned only as a duration object.}
}
\value{
A grouped tibble data frame with the following columns:
\itemize{
\item \code{group_by} names, e.g. \code{deploymentID}, \code{latitude}, \code{longitude} and
\code{locationName}.
\item \code{group_time_by} name if provided, e.g. \code{month}. It contains the first
date of the time interval, e.g. the first day of the month.
\item \code{effort_duration}: A duration object (duration is a class from lubridate
package). Duration is always recorded as a fixed number of seconds. See
\code{\link[lubridate:duration]{lubridate::duration()}}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Gets the effort for each deployment and a specific time interval such as day,
week, month or year.

It is deprecated as of camtraptor 1.0.0. Please use \code{\link[=summarize_deployments]{summarize_deployments()}}
instead.
}
\examples{
x <- example_dataset()

# Effort for each deployment over the entire duration of the project
get_custom_effort(x)

# Effort at monthly interval
get_custom_effort(x, group_by = "month")

# You can afterwards calculate the total effort over all deployments
library(dplyr)
get_custom_effort(x, group_by = "year") \%>\%
  filter(effort_duration > 0) \%>\%
  group_by(year) \%>\% 
  summarise(
    deploymentIDs = list(deploymentID),
    ndep = n_distinct(deploymentID),
    effort_duration = sum(effort_duration)
  )
}
\seealso{
Other deprecated exploration functions:
\code{\link[=get_effort]{get_effort()}},
\code{\link[=get_n_individuals]{get_n_individuals()}},
\code{\link[=get_n_obs]{get_n_obs()}},
\code{\link[=get_n_species]{get_n_species()}},
\code{\link[=get_rai]{get_rai()}},
\code{\link[=get_rai_individuals]{get_rai_individuals()}}
}
\concept{deprecated exploration functions}
