% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_effort.R
\name{get_custom_effort}
\alias{get_custom_effort}
\title{Get custom effort}
\usage{
get_custom_effort(
  datapkg,
  ...,
  start = NULL,
  end = NULL,
  group_by = NULL,
  unit = NULL
)
}
\arguments{
\item{datapkg}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates}

\item{start}{Start date. Default: \code{NULL}. If \code{NULL} the earliest start date
among all deployments is used.}

\item{end}{End date. Default: \code{NULL}. If \code{NULL} the latest end date among all
deployments is used.}

\item{group_by}{Character, one of \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"}. The
effort is calculated at the interval rate defined in \code{group_by}. Default:
\code{NULL}: no grouping, i.e. the entire interval from \code{start} to \code{end} is
taken into account as a whole.}

\item{unit}{Time unit to use while returning custom effort
(duration). One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
\item \code{NULL} (default) duration objects (e.g. 2594308s (~4.29 weeks))
}}
}
\value{
A tibble (data.frame) with one column:
\itemize{
\item \code{effort}: duration object (duration is a class from lubridate package) if \code{unit} is \code{NULL}, or a number if \code{unit} is not \code{NULL}.
\item \code{effort_unit}: Character specifying the effort_unit.
}
}
\description{
Function to get custom effort (deployment duration) for a custom time windows
and a specific time interval such as day or month. The custom effort is also
calculated over all deployments, although filtering predicates can be applied
as well.
}
\examples{
# a global effort over the entire duration of the project (datapackage)
get_custom_effort(mica)

# effort expressed as days
get_custom_effort(mica, unit = "day")

# effort from a specific start to a specific end
get_custom_effort(
  mica,
  start = as.Date("2019-12-15"), # or lubridate::as_date("2019-12-15")
  end = as.Date("2021-01-10")
)

# effort at weekly interval
get_custom_effort(
  mica,
  start = as.Date("2019-12-15"),
  end = as.Date("2021-01-10"),
  group_by = "week"
)
}
\seealso{
Other get_functions: 
\code{\link{filter_predicate}},
\code{\link{get_effort}()}
}
\concept{get_functions}
