% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_dep_no_obs}
\alias{get_dep_no_obs}
\title{Get deployments with no observations}
\usage{
get_dep_no_obs(package = NULL, ..., datapkg = lifecycle::deprecated())
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{Filter predicates for filtering on deployments}
}
\value{
Tibble data.frame with deployments not linked to any observations.
}
\description{
Return subset of deployments without observations. A message is also returned
to list the ID of such deployments.
}
\examples{
get_dep_no_obs(mica)
}
