% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_dep_no_obs}
\alias{get_dep_no_obs}
\title{Get deployments with no observations}
\usage{
get_dep_no_obs(package = NULL, ..., datapkg = lifecycle::deprecated())
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{Filter predicates for filtering on deployments}

\item{datapkg}{Deprecated. Use \code{package} instead.}
}
\value{
Tibble data.frame with deployments not linked to any observations.
}
\description{
Return subset of deployments without observations. A message is also returned
to list the ID of such deployments.
}
\examples{
get_dep_no_obs(mica)
}
\seealso{
Other exploration functions: 
\code{\link{get_cam_op}()},
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_n_species}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_rai}()},
\code{\link{get_record_table}()},
\code{\link{get_scientific_name}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
