% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_dep_no_obs}
\alias{get_dep_no_obs}
\title{Get deployments with no observations}
\usage{
get_dep_no_obs(datapkg, ...)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{...}{filter predicates for filtering on deployments}
}
\value{
a tibble (data.frame) with deployments not linked to any observations
}
\description{
Return subset of deployments without observations. A message is also returned
to list the ID of such deployments.
}
