% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detection_history.R
\name{get_detection_history}
\alias{get_detection_history}
\title{Get the detection history of a species}
\usage{
get_detection_history(
  recordTable,
  camOp,
  species,
  output,
  occasionLength = 1,
  minActiveDaysPerOccasion = NULL,
  maxNumberDays = NULL,
  day1 = "station",
  buffer = NULL,
  unmarkedMultFrameInput = FALSE
)
}
\arguments{
\item{recordTable}{A record table created by \code{get_record_table()}.}

\item{camOp}{A camera operation matrix created by \code{get_cam_op()}.}

\item{species}{Character. The species name.}

\item{output}{Character. The type of output. Choose one of: \code{"binary"},
\code{"n_observations"}, \code{"n_individuals"}.}

\item{occasionLength}{Integer. The length of the occasions in days. No
decimals allowed. Default: \code{1}.}

\item{minActiveDaysPerOccasion}{Integer. Minimum number of active trap days
for occasions to be included. Default: \code{NULL}. If used, it must be smaller
than or equal to \code{occasionLength}.}

\item{maxNumberDays}{Integer. Maximum number of trap days per station.
Default: \code{NULL}. If used, it must be greater than or equal to
\code{occasionLength}.}

\item{day1}{Character. Day occasions should begin: station setup date
(\code{"station"}) or a specific date (e.g. \code{"2015-12-31"}). For multi-season
detection history (\code{unmarkedMultFrameInput} = \code{TRUE}), only \code{day1} =
\code{"station"} is allowed. Default: \code{"station"}.}

\item{buffer}{Integer. It makes the first occasion begin a number of days
after station setup. \code{buffer} can be used only in combination with \code{day1} =
\code{"station"}. Default: \code{NULL}. A warning is returned if some records are
removed because taken during the buffer period.}

\item{unmarkedMultFrameInput}{Logical. If \code{TRUE}, the function will return
the input for multi-season occupancy models in unmarked
R package (argument \code{y} in \code{\link[unmarked:unmarkedMultFrame]{unmarked::unmarkedMultFrame()}}).
Default: \code{FALSE}.}
}
\value{
A list with three elements:
\itemize{
\item \code{detection_history}: the detection history matrix
\item \code{effort}: the effort matrix
\item \code{dates}: the dates matrix
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

This function is superseded because camtrapR supports reading Camera Trap
Data Packages. Use \code{\link[camtrapR:readCamtrapDP]{camtrapR::readCamtrapDP()}} and
\code{\link[camtrapR:detectionHistory]{camtrapR::detectionHistory()}} instead.

Creates species detection matrices. See
\code{vignette("detection-history-matrix")} for more details on how to use
this function.
}
\details{
If the camera operation matrix (\code{camOp}) was created for a multi-season study
(via argument \code{session_col} in \code{get_cam_op()}), the session will be detected
automatically. You can then set \code{unmarkedMultFrameInput} = \code{TRUE} to generate
a multi-season detection history. Each row corresponds to a site, and the
columns are in season-major, occasion-minor order, e.g. \code{o1_SESS_A},
\code{o2_SESS_A}, \code{o1_SESS_B}, \code{o2_SESS_B}, etc.
}
\examples{
library(dplyr)
library(stringr)

x <- example_dataset()
camOp <- get_cam_op(x)
recordTable <- get_record_table(x)
# More observations of the same species on the same day at the same station
# are left.
recordTable_mulitple <- get_record_table(
  x,
  removeDuplicateRecords = FALSE
)

# Binary output
get_detection_history(
  recordTable,
  camOp,
  species = "Anas platyrhynchos",
  output = "binary"
)

# Number of observations output: same as binary with default
# `get_record_table(x)`
get_detection_history(
 recordTable,
 camOp,
 species = "Anas platyrhynchos",
 output = "n_observations"
)

# Number of observations output: more than 1 if more than 1 observation of
# the species on the same day at the same station
get_detection_history(
 recordTable_mulitple,
 camOp,
 species = "Anas platyrhynchos",
 output = "n_observations"
)

# Number of individuals output
get_detection_history(
 recordTable,
 camOp,
 species = "Anas platyrhynchos",
 output = "n_individuals"
)

# Occasion length of 7 days
get_detection_history(
 recordTable,
 camOp,
 species = "Anas platyrhynchos",
 output = "n_individuals",
 occasionLength = 7
)

# Use a `minActiveDaysPerOccasion` of 5 days
get_detection_history(
 recordTable,
 camOp,
 species = "Anas platyrhynchos",
 output = "n_individuals",
 occasionLength = 7,
 minActiveDaysPerOccasion = 5
)

# Use a `maxNumberDays` of 5 days
get_detection_history(
 recordTable,
 camOp,
 species = "Anas platyrhynchos",
 output = "n_individuals",
 maxNumberDays = 5
)

# Specify start date via `day1`
get_detection_history(
  recordTable,
  camOp,
  species = "Anas platyrhynchos",
  output = "binary",
  day1 = "2020-06-22"
)

# Use a `buffer` of 2 days
get_detection_history(
 recordTable,
 camOp,
 species = "Anas platyrhynchos",
 output = "n_individuals",
 buffer = 2
)

# Multi-season detection history
x_sessions <- x
deployments(x_sessions) <- deployments(x_sessions) \%>\%
  mutate(session = ifelse(
    str_starts(.data$locationName, "B_DL_"),
      "after2020",
      "before2020"
  )
)
camOp_sessions <- get_cam_op(
  x_sessions,
  session_col = "session"
)
recordTable_sessions <- get_record_table(x_sessions)

# Create a multi-season detection history
get_detection_history(
  recordTable_sessions,
  camOp_sessions,
  species = "Anas platyrhynchos",
  output = "n_individuals",
  unmarkedMultFrameInput = TRUE
)
}
\seealso{
Other deprecated camtrapR-derived functions:
\code{\link[=get_cam_op]{get_cam_op()}},
\code{\link[=get_record_table]{get_record_table()}}
}
\concept{deprecated camtrapR-derived functions}
