% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detection_history.R
\name{get_detection_history}
\alias{get_detection_history}
\title{Get the detection history of a species}
\usage{
get_detection_history(cam_op, rec_table, species, output, occasion_length = 1)
}
\arguments{
\item{cam_op}{A matrix with camera operation data. Rows represent camera stations and columns represent occasions. The matrix should contain the number of days that each station was active on each occasion.}

\item{rec_table}{A data frame with the camera trap records. The data frame should contain the columns 'Station', 'Date', 'Species', 'n', and 'n_ind'. 'Station' is the camera station ID, 'Date' is the date of the record, 'Species' is the species name, 'n' is the number of observations, and 'n_ind' is the number of individuals detected.}

\item{species}{The species name.}

\item{output}{The type of output. Choose one of: 'binary', 'n_observations', 'n_individuals'.}

\item{occasion_length}{The length of the occasions in days. Default: 1.}
}
\value{
A list with three elements: 'detection_history' is the detection history matrix, 'effort' is the effort matrix, and 'dates' is the dates matrix.
}
\description{
This function creates a detection history matrix for a species from a camera trap dataset.
The detection history matrix is a binary matrix where rows represent camera stations and columns represent occasions.
The matrix is filled with 1s and 0s, where 1 indicates that the species was detected at a station on a given occasion and 0 indicates that the species was not detected.
The function also returns the effort matrix, which contains the number of days that each station was active on each occasion, and the dates matrix, which contains the dates of the occasions.
}
\examples{
cam_op <- get_cam_op(mica)
rec_table <- get_record_table(mica)
get_detection_history(cam_op, rec_table, "Anas platyrhynchos", "binary")
}
