% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{get_effort}
\alias{get_effort}
\title{Get effort}
\usage{
get_effort(datapkg)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{multimedia}
}

and a list with metadata: \code{datapackage}}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deployment_id} deployment unique identifier
\item \code{effort}: a duration object (duration is a class from lubridate package)
}
}
\description{
Function to get the effort (deployment duration) per deployment.
}
\examples{
get_effort(camtrapdp)

}
