% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{get_effort}
\alias{get_effort}
\title{Get effort}
\usage{
get_effort(
  package = NULL,
  ...,
  unit = "hour",
  datapkg = lifecycle::deprecated()
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates}

\item{unit}{Time unit to use while returning deployment effort
(duration). One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
}}

\item{datapkg}{Deprecated. Use \code{package} instead.}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deploymentID}: deployment unique identifier.
\item \code{effort}: effort expressed in the unit passed by argument \code{unit}.
\item \code{unit}: the unit used to express the effort. One of the values available
for argument \code{unit}.
\item \code{effort_duration}: a duration object (duration is a class from lubridate package).
}
}
\description{
Function to get the effort (deployment duration) per deployment.
}
\examples{
# efforts expressed in hours
get_effort(mica)

# effort expressed as days
get_effort(mica, unit = "day")

}
\seealso{
Other get_functions: 
\code{\link{filter_predicate}},
\code{\link{get_cam_op}()},
\code{\link{get_custom_effort}()},
\code{\link{get_record_table}()}
}
\concept{get_functions}
