% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{get_effort}
\alias{get_effort}
\title{Get effort}
\usage{
get_effort(datapkg, ..., unit = NULL)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates}

\item{unit}{time unit to use while returning deployment effort
(duration). One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
\item \code{NULL} (default) duration objects (e.g. 2594308s (~4.29 weeks))
}}
}
\value{
a tibble (data.frame) with the following columns: - \code{deploymentID}
deployment unique identifier - \code{effort}: a duration object (duration is a
class from lubridate package) - \code{unit}: the unit used to express the
effort. One of the values available for argument \code{unit}, except \code{NULL}. If
\code{NULL}, the returned column \code{unit} is equal to \code{"Duration"}.
}
\description{
Function to get the effort (deployment duration) per deployment.
}
\examples{
# efforts expressed as Durations
get_effort(mica)

# effort expressed as days
get_effort(mica, unit = "day")

}
\seealso{
Other get_functions: 
\code{\link{filter_predicate}},
\code{\link{get_custom_effort}()}
}
\concept{get_functions}
