% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{get_effort}
\alias{get_effort}
\title{Get effort}
\usage{
get_effort(x, unit = "hour")
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{unit}{Deprecated and not supported anymore.}
}
\value{
A tibble data frame with following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{effort_duration}: A duration object (duration is a class from lubridate
package).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is deprecated. Please use \code{\link[=summarize_deployments]{summarize_deployments()}} instead.

This function returns the effort (deployment duration) per deployment.
}
\examples{
x <- example_dataset()
get_effort(x)
}
\seealso{
Other exploration functions: 
\code{\link{get_custom_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{n_species}()},
\code{\link{summarize_deployments}()}
}
\concept{exploration functions}
