% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{get_effort}
\alias{get_effort}
\title{Get effort}
\usage{
get_effort(datapkg, ..., unit = NULL)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates}

\item{unit}{time unit to use while returning deployment effort
(duration). One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
\item \code{NULL} (default) duration objects (e.g. 2594308s (~4.29 weeks))
}}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deployment_id} deployment unique identifier
\item \code{effort}: a duration object (duration is a class from lubridate package)
}
}
\description{
Function to get the effort (deployment duration) per deployment.
}
\examples{
# efforts expressed as Durations
get_effort(camtrapdp)

# effort expressed as days
get_effort(camtrapdp, unit = "day")

}
\seealso{
Other get_functions: 
\code{\link{filter_predicate}}
}
\concept{get_functions}
