% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{get_effort}
\alias{get_effort}
\title{Get effort}
\usage{
get_effort(
  package = NULL,
  ...,
  unit = "hour",
  datapkg = lifecycle::deprecated()
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{...}{filter predicates}

\item{unit}{Time unit to use while returning deployment effort (duration).
One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
}}

\item{datapkg}{Deprecated.
Use \code{package} instead.}
}
\value{
A tibble data frame with following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{effort}: Effort expressed in the unit passed by parameter \code{unit}.
\item \code{unit}: The unit used to express the effort.
One of the values available for parameter \code{unit}.
\item \code{effort_duration}: A duration object (duration is a class from lubridate
package).
}
}
\description{
Gets the effort (deployment duration) per deployment.
}
\examples{
# Efforts expressed in hours
get_effort(mica)

# Effort expressed as days
get_effort(mica, unit = "day")
}
\seealso{
Other exploration functions: 
\code{\link{get_cam_op}()},
\code{\link{get_custom_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_n_species}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_record_table}()},
\code{\link{get_scientific_name}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
