% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_individuals.R
\name{get_n_individuals}
\alias{get_n_individuals}
\title{Get number of individuals for each deployment}
\usage{
get_n_individuals(datapkg, ..., species = "all", sex = NULL, life_stage = NULL)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates for filtering on deployments}

\item{species}{a character with scientific names or common names (case
insensitive). If "all", default, all scientific names are automatically
selected. If \code{NULL} all observations of all species are taken into account}

\item{sex}{a character defining the sex class to filter on, e.g. \code{"female"}
or \code{c("male", "unknown")}.  If \code{NULL}, default, all observations of all
sex classes are taken into account.}

\item{life_stage}{a character vector defining the life stage class to filter on, e.g.
\code{"adult"} or \code{c("subadult", "adult")}. If \code{NULL}, default, all observations
of all life stage classes are taken into account.}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deploymentID}:  deployment unique identifier
\item \code{scientificName}: scientific name of the species. This column is omitted
if argument \code{species} = NULL
\item \code{n}: (integer) number of individuals
}
}
\description{
Function to get the number of individuals (of a subset of species) per
deployment. The number of observed individuals is stored in field \code{count} of
\code{observations}.
}
\examples{

# get number of observations for each species
get_n_individuals(mica)

# get number of obs of all species, not identified individuals as well
get_n_individuals(mica, species = NULL)

# get number of observations of Anas platyrhynchos
get_n_individuals(mica, species = "Anas platyrhynchos")

# get number of observations of eurasian beaver (vernacular name)
get_n_individuals(mica, species = "eurasian beaver")

# mix scientific and vernacular names
get_n_individuals(mica, species = c("Anas platyrhynchos", "eurasian beaver"))

# case insensitive
get_n_individuals(mica, species = "AnAS PLatyrhyncHOS")
get_n_individuals(mica, species = "eurasian BEAVER")

# specify life stage
get_n_individuals(mica, life_stage = "adult")

# specify sex
get_n_individuals(mica, sex = "female")

# specify both sex and life stage
get_n_individuals(mica, sex = "unknown", life_stage = "adult")

# applying filter(s), e.g. deployments with latitude >= 51.18
get_n_individuals(mica, pred_gte("latitude", 51.18))

}
