% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_individuals.R
\name{get_n_individuals}
\alias{get_n_individuals}
\title{Get number of individuals for each deployment}
\usage{
get_n_individuals(x)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{scientificName}: Scientific name of the species. \code{NA} for not identified individuals.
\item \code{n}: Number of individuals.
}
}
\description{
Gets the number of individuals (of a subset of species) per deployment.
The number of observed individuals is stored in column \code{count} of
\code{observations}.
}
\examples{
x <- example_dataset()

# Get number of individuals for each species
get_n_individuals(x)

# Use `filter_observations()` to filter on  scientific name
x \%>\%
  filter_observations(
    scientificName \%in\% c("Anas platyrhynchos", "Vulpes vulpes")
  ) \%>\%
  get_n_individuals()

# Use `filter_observations()` to filter on vernacular name
x \%>\%
  filter_observations(taxon.vernacularNames.eng == "mallard") \%>\%
  get_n_individuals()

# Get number of individuals per deployment
x \%>\%
  get_n_individuals() \%>\%
  dplyr::group_by(.data$deploymentID) \%>\%
  dplyr::summarise(n = sum(.data$n)) \%>\%
  dplyr::ungroup()
}
\seealso{
Other exploration functions: 
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{n_species}()}
}
\concept{exploration functions}
