% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_individuals.R
\name{get_n_individuals}
\alias{get_n_individuals}
\title{Get number of individuals for each deployment}
\usage{
get_n_individuals(
  package = NULL,
  ...,
  species = "all",
  sex = NULL,
  life_stage = NULL,
  datapkg = lifecycle::deprecated()
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates for filtering on deployments}

\item{species}{Character with scientific names or common names (case
insensitive).
If \code{"all"} (default) all scientific names are automatically selected.
If \code{NULL} all observations of all species are taken into account.}

\item{sex}{Character defining the sex class to filter on, e.g. \code{"female"}
or \code{c("male", "unknown")}.
If \code{NULL} (default) all observations of all sex classes are taken into
account.}

\item{life_stage}{Character vector defining the life stage class to filter
on, e.g. \code{"adult"} or \code{c("subadult", "adult")}.
If \code{NULL} (default) all observations of all life stage classes are taken
into account.}

\item{datapkg}{Deprecated.
Use \code{package} instead.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{scientificName}: Scientific name of the species.
This column is omitted if parameter \code{species = NULL}.
\item \code{n}: Number of individuals.
}
}
\description{
Gets the number of individuals (of a subset of species) per deployment.
The number of observed individuals is stored in field \code{count} of
\code{observations}.
}
\examples{
# Get number of observations for each species
get_n_individuals(mica)

# Get number of obs of all species, not identified individuals as well
get_n_individuals(mica, species = NULL)

# Get number of observations of Anas platyrhynchos
get_n_individuals(mica, species = "Anas platyrhynchos")

# Get number of observations of eurasian beaver (vernacular name)
get_n_individuals(mica, species = "eurasian beaver")

# Mix scientific and vernacular names
get_n_individuals(mica, species = c("Anas platyrhynchos", "eurasian beaver"))

# Case insensitive
get_n_individuals(mica, species = "AnAS PLatyrhyncHOS")
get_n_individuals(mica, species = "eurasian BEAVER")

# Specify life stage
get_n_individuals(mica, life_stage = "adult")

# Specify sex
get_n_individuals(mica, sex = "female")

# Specify both sex and life stage
get_n_individuals(mica, sex = "unknown", life_stage = "adult")

# Apply filter(s), e.g. deployments with latitude >= 51.18
get_n_individuals(mica, pred_gte("latitude", 51.18))
}
\seealso{
Other exploration functions: 
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_obs}()},
\code{\link{get_n_species}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_scientific_name}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
