% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_obs.R
\name{get_n_obs}
\alias{get_n_obs}
\title{Get number of observations for each deployment}
\usage{
get_n_obs(deployments, observations)
}
\arguments{
\item{deployments}{a tibble (data.frame) containing deployments}

\item{observations}{a tibble (data.frame) containing observations}
}
\value{
a tibble (data.frame) with the following columns: - \code{deployment_id}
deployment unique identifier - \code{n}: (integer) number of observations
}
\description{
Function to get the number of observations per deployment.
}
