% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_obs.R
\name{get_n_obs}
\alias{get_n_obs}
\title{Get number of observations for each deployment}
\usage{
get_n_obs(x)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{scientificName}: Scientific name of the species.
\item \code{n}: Number of observations.
}
}
\description{
Gets the number of observations per deployment.
The number of observations is defined as the number of distinct sequences
(\code{sequenceID}).
}
\examples{
x <- example_dataset()

# Get number of individuals for each species
get_n_obs(x)

# Use `filter_observations()` to filter on  scientific name
x \%>\%
  filter_observations(
    scientificName \%in\% c("Anas platyrhynchos", "Vulpes vulpes")
  ) \%>\%
  get_n_obs()

# Use `filter_observations()` to filter on vernacular name
x \%>\%
  filter_observations(taxon.vernacularNames.eng == "mallard") \%>\%
  get_n_obs()

# Get number of individuals per deployment
x \%>\%
  get_n_obs() \%>\%
  dplyr::group_by(.data$deploymentID) \%>\%
  dplyr::summarise(n = sum(.data$n)) \%>\%
  dplyr::ungroup()
}
\seealso{
Other exploration functions: 
\code{\link{get_n_individuals}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{n_species}()},
\code{\link{summarize_deployments}()}
}
\concept{exploration functions}
