% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_obs.R
\name{get_n_obs}
\alias{get_n_obs}
\title{Get number of observations for each deployment}
\usage{
get_n_obs(datapkg, ..., species = "all", sex = NULL, age = NULL)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates for filtering on deployments}

\item{species}{a character with scientific names or common names (case
insensitive). If "all", default, all scientific names are automatically
selected. If \code{NULL} all observations of all species are taken into account}

\item{sex}{a character defining the sex class to filter on, e.g. \code{"female"}
or \code{c("male", "undefined")}.  If \code{NULL}, default, all observations of all
sex classes are taken into account.}

\item{age}{a character vector defining the age class to filter on, e.g.
\code{"adult"} or \code{c("subadult", "adult")}. If \code{NULL}, default, all observations
of all age classes are taken into account.}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deployment_id}:  deployment unique identifier
\item \code{scientific_name}: scientific name of the species. This column is omitted
if argument \code{species} = NULL
\item \code{n}: (integer) number of observations
}
}
\description{
Function to get the number of observations (of a subset of species) per
deployment. The number of observations is defined as the number of distinct
sequences (\code{sequence_id}).
}
\examples{

# get number of observations for each species
get_n_obs(camtrapdp)

# get number of obs of all species, not identified individuals as well
get_n_obs(camtrapdp, species = NULL)

# get number of observations of Gallinula chloropus
get_n_obs(camtrapdp, species = "Gallinula chloropus")

# get number of observations of Common Moorhen
get_n_obs(camtrapdp, species = "Common Moorhen")

# case insensitive
get_n_obs(camtrapdp, species = "galliNULa CHloropUs")
get_n_obs(camtrapdp, species = "cOmmon moorhEn")

# specify age
get_n_obs(camtrapdp, age = "adult")

# specify sex
get_n_obs(camtrapdp, sex = "female")

# specify both sex and age
get_n_obs(camtrapdp, sex = "undefined", age = "adult")

# applying filter(s), e.g. deployments with latitude >= 51.28
get_n_obs(camtrapdp, pred_gte("latitude", 51.28))

}
