% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_obs.R
\name{get_n_obs}
\alias{get_n_obs}
\title{Get number of observations for each deployment}
\usage{
get_n_obs(datapkg)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{multimedia}
}

and a list with metadata: \code{datapackage}}
}
\value{
a tibble (data.frame) with the following columns: - \code{deployment_id}
deployment unique identifier - \code{n}: (integer) number of observations
}
\description{
Function to get the number of observations per deployment.
}
