% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_obs.R
\name{get_n_obs}
\alias{get_n_obs}
\title{Get number of observations for each deployment}
\usage{
get_n_obs(package = NULL, ..., species = "all", sex = NULL, life_stage = NULL)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{Filter predicates for filtering on deployments}

\item{species}{Character with scientific names or common names (case
insensitive).
If \code{"all"} (default) all scientific names are automatically selected.
If \code{NULL} all observations of all species are taken into account.}

\item{sex}{Character defining the sex class to filter on, e.g. \code{"female"} or
\code{c("male", "unknown")}.
If \code{NULL} (default) all observations of all sex classes are taken into
account.}

\item{life_stage}{Character vector defining the life stage class to filter
on, e.g. \code{"adult"} or \code{c("subadult", "adult")}.
If \code{NULL} (default) all observations of all life stage classes are taken
into account.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{scientificName}: Scientific name of the species.
This column is omitted if parameter \code{species = NULL}.
\item \code{n}: Number of observations.
}
}
\description{
Gets the number of observations (of a subset of species) per deployment.
The number of observations is defined as the number of distinct sequences
(\code{sequenceID}).
}
\examples{
# Get number of observations for each species
get_n_obs(mica)

# Get number of obs of all species, not identified individuals as well
get_n_obs(mica, species = NULL)

# Get number of observations of Anas platyrhynchos (scientific name)
get_n_obs(mica, species = "Anas platyrhynchos")

# Get number of observations of mallard (vernacular names)
get_n_obs(mica, species = "mallard")

# Case insensitive
get_n_obs(mica, species = "Anas plaTYrhYnchoS")

# Specify life stage
get_n_obs(mica, life_stage = "subadult")

# Specify sex
get_n_obs(mica, sex = "female")

# Specify both sex and life stage
get_n_obs(mica, sex = NULL, life_stage = "adult")

# Applying filter(s), e.g. deployments with latitude >= 51.18
get_n_obs(mica, pred_gte("latitude", 51.18))
}
\seealso{
Other exploration functions: 
\code{\link{get_cam_op}()},
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_species}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_rai}()},
\code{\link{get_record_table}()},
\code{\link{get_scientific_name}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
