% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_obs.R
\name{get_n_obs}
\alias{get_n_obs}
\title{Get number of observations for each deployment}
\usage{
get_n_obs(datapkg, species = "all")
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations} 2. \code{deployments} 3. \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{species}{a character with scientific names or common names (case
insensitive). If "all" (default), all scientific names are automatically
selected. If \code{NULL} all observations of all species are taken into account}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deployment_id}:  deployment unique identifier
\item \code{scientific_name}: scientific name of the species. This column is omitted
if argument \code{species} = NULL
\item \code{n}: (integer) number of observations
}
}
\description{
Function to get the number of observations (of a subset of species) per
deployment.
}
\examples{

# get number of observations for each species
get_n_obs(camtrapdp, species = "all")

# get number of observations of all species (not identified individuals as well)
get_n_obs(camtrapdp)

# get number of observations of Gallinula chloropus
get_n_obs(camtrapdp, species = "Gallinula chloropus")

# get number of observations of Common Moorhen
get_n_obs(camtrapdp, species = "Common Moorhen")

# case insensitive
get_n_obs(camtrapdp, species = "cOmmon moorhEn")

}
