% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_obs.R
\name{get_n_obs}
\alias{get_n_obs}
\title{Get number of observations for each deployment}
\usage{
get_n_obs(x)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
A tibble data frame with following columns:
\itemize{
\item \code{deploymentID}: deployment unique identifier.
\item \code{n_observations}: intger vector with the number of observations.
\item \code{sum_count}: integer vector with the sum of the count of individuals.
\item \code{rai}: numeric vector with the relative abundance index (RAI) based on
number of observations.
\item \code{rai_individuals}: numeric vector with the RAI based on the count of
individuals.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Gets the number of observations per deployment.

It is deprecated as of camtraptor 1.0.0. Please use \code{\link[=summarize_observations]{summarize_observations()}}
instead.
}
