% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_species.R
\name{get_n_species}
\alias{get_n_species}
\title{Get number of identified species for each deployment}
\usage{
get_n_species(package = NULL, ..., datapkg = lifecycle::deprecated())
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{Filter predicates for filtering on deployments}

\item{datapkg}{Deprecated. Use \code{package} instead.}
}
\value{
Tibble data.frame with the following columns:
\itemize{
\item \code{deploymentID}: deployment unique identifier
\item \code{n}: (integer) number of observed and identified species
}
}
\description{
Function to get the number of identified species per deployment.
}
\examples{

# get number of species
get_n_species(mica)

# get number of species for deployments with latitude >= 51.18
get_n_species(mica, pred_gte("latitude", 51.18))

}
