% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_species.R
\name{get_n_species}
\alias{get_n_species}
\title{Get number of identified species for each deployment}
\usage{
get_n_species(x, ...)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{...}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} filter predicates for filtering
on deployments are not supported anymore and an error is returned. Anything
else throws a deprecation warning and will be ignored. Please use
\code{filter_deployments()} to filter on deployments.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{n}: Number of observed and identified species.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

It is deprecated as of camtraptor 1.0.0. Please use \code{\link[=summarize_observations]{summarize_observations()}}
instead.
}
\examples{
x <- example_dataset()
# Get number of species
get_n_species(x)
}
