% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_species.R
\name{get_n_species}
\alias{get_n_species}
\title{Get number of identified species for each deployment}
\usage{
get_n_species(datapkg, ...)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates for filtering on deployments}
}
\value{
a tibble (data.frame) with the following columns: - \code{deployment_id}
deployment unique identifier - \code{n}: (integer) number of observed and
identified species
}
\description{
Function to get the number of identified species per deployment.
}
\examples{

# get number of species
get_n_species(camtrapdp)

# get number of species for deployments with latitude >= 51.28
get_n_species(camtrapdp, pred_gte("latitude", 51.28))

}
