% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_species.R
\name{get_n_species}
\alias{get_n_species}
\title{Get number of identified species for each deployment}
\usage{
get_n_species(x)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{n}: Number of observed and identified species.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=n_species]{n_species()}} instead.
}
\details{
Only event-based observations (\code{observationLevel} = \code{event}) are taken into account.
}
\examples{
x <- example_dataset()

# Get number of species
n_species(x)
}
\seealso{
Other exploration functions: 
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{summarize_deployments}()},
\code{\link{summarize_observations}()}
}
