% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_species.R
\name{get_n_species}
\alias{get_n_species}
\title{Get number of identified species for each deployment}
\usage{
get_n_species(package = NULL, ...)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{Filter predicates for filtering on deployments.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{n}: Number of observed and identified species.
}
}
\description{
Gets the number of identified species per deployment.
}
\examples{
# Get number of species
get_n_species(mica)

# Get number of species for deployments with latitude >= 51.18
get_n_species(mica, pred_gte("latitude", 51.18))
}
\seealso{
Other exploration functions: 
\code{\link{get_cam_op}()},
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_rai}()},
\code{\link{get_record_table}()},
\code{\link{get_scientific_name}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
