% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rai.R
\name{get_rai}
\alias{get_rai}
\title{Get Relative Abundance Index (RAI)}
\usage{
get_rai(datapkg, species)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{species}{a character with scientific name}
}
\value{
a tibble (data.frame) with the following columns: - \code{deployment_id}
deployment unique identifier - \code{rai}: relative abundance index
}
\description{
Function to get the RAI (Relative Abundance Index) per deployment.
}
