% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rai.R
\name{get_rai}
\alias{get_rai}
\title{Get Relative Abundance Index (RAI)}
\usage{
get_rai(
  package = NULL,
  ...,
  species = "all",
  sex = NULL,
  life_stage = NULL,
  datapkg = lifecycle::deprecated()
)
}
\arguments{
\item{package}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{...}{filter predicates for filtering on deployments}

\item{species}{a character with scientific names or common names (case
insensitive). If "all" (default), all scientific names are automatically
selected}

\item{sex}{a character defining the sex class to filter on, e.g. \code{"female"}
or \code{c("male", "unknown")}.  If \code{NULL}, default, all observations of all
sex classes are taken into account.}

\item{life_stage}{a character vector defining the life stage class to filter on, e.g.
\code{"adult"} or \code{c("subadult", "adult")}. If \code{NULL}, default, all observations
of all life stage classes are taken into account.}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deploymentID} deployment unique identifier
\item \code{scientificName} scientific name
\item \code{rai}: relative abundance index
}
}
\description{
Function to get the RAI (Relative Abundance Index) per
deployment. The RAI is normalized using 100 days deployment activity.
}
\examples{
# calculate RAI for all species
get_rai(mica) # species = "all" by default, so equivalent of
get_rai(mica, species = "all")

# selected species
get_rai(mica, species = c("Anas platyrhynchos", "Martes foina"))

# with vernacular names, even mixing languages
get_rai(mica, species = c("mallard", "steenmarter"))

# mixed scientific and vernacular names
get_rai(mica, species = c("Anas platyrhynchos", "steenmarter"))

# species argument is case insensitive
get_rai(mica, species = c("ANAS plAtyRhynChOS"))

# specify sex
get_rai(mica, sex = "female")
get_rai(mica, sex = c("female", "unknown"))

# specify life stage
get_rai(mica, life_stage = "adult")
get_rai(mica, life_stage = c("adult", "subadult"))

# apply filter(s): deployments with latitude >= 51.18
get_rai(mica, pred_gte("latitude", 51.18))

}
\seealso{
Other RAI functions: 
\code{\link{get_rai_individuals}()}
}
\concept{RAI functions}
