% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rai.R
\name{get_rai}
\alias{get_rai}
\title{Get Relative Abundance Index (RAI)}
\usage{
get_rai(datapkg, ..., species = "all")
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{...}{filter predicates for filtering on deployments}

\item{species}{a character with scientific names or common names (case
insensitive). If "all" (default), all scientific names are automatically
selected}
}
\value{
a tibble (data.frame) with the following columns:
\itemize{
\item \code{deployment_id} deployment unique identifier
\item \code{scientific_name} scientific name
\item \code{rai}: relative abundance index
}
}
\description{
Function to get the RAI (Relative Abundance Index) per deployment.
}
\examples{
# all species
get_rai(camtrapdp, species = "all")

# selected species
get_rai(camtrapdp, species = c("Anas platyrhynchos", "Rattus norvegicus"))

# with common names
get_rai(camtrapdp, species = c("Mallard", "norway rat"))

# mixed scientific and vernacular names
get_rai(camtrapdp, species = c("Anas platyrhynchos", "norway rat"))

# species argument is case insensitive
get_rai(camtrapdp, species = c("ANAS plAtyRhynChOS"))

# apply filtering: deployments with latitude >= 51.28
get_rai(camtrapdp, species = "all", pred_gte("latitude", 51.28))

}
