% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rai.R
\name{get_rai}
\alias{get_rai}
\title{Get Relative Abundance Index (RAI)}
\usage{
get_rai(x, species = "all")
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{species}{Character with scientific names or common names (case
insensitive). If \code{"all"} (default) all scientific names are automatically
selected.}
}
\value{
A tibble data frame with the following columns: - \code{deploymentID}:
Deployment unique identifier. - \code{scientificName}: Scientific name. - \code{rai}:
Relative abundance index.
}
\description{
Gets the RAI (Relative Abundance Index) per deployment. The RAI is normalized
using 100 days deployment activity. In other words:
\code{RAI = 100 * (n/effort)} where \code{n} is the number of observations as
calculated via \code{get_n_obs()} and \code{effort} is the effort in days as calculated
via \code{get_effort()}.
}
\examples{
# Calculate RAI for all species
get_rai(x) # species = "all" by default, so equivalent of
get_rai(x, species = "all")

# Selected species
get_rai(x, species = c("Anas platyrhynchos", "Martes foina"))

# With vernacular names, even mixing languages
get_rai(x, species = c("mallard", "steenmarter"))

# Mixed scientific and vernacular names
get_rai(x, species = c("Anas platyrhynchos", "steenmarter"))

# Species parameter is case insensitive
get_rai(x, species = c("ANAS plAtyRhynChOS"))

# Use `filter_observations()` to filter on life stage
x \%>\%
  filter_observations(lifeStage == "adult") \%>\%
  get_rai()

# Use `filter_observations()` to filter on sex
x \%>\%
  filter_observations(sex == "female") \%>\%
  get_rai()
}
\seealso{
Other exploration functions: 
\code{\link{get_rai_individuals}()},
\code{\link{n_species}()},
\code{\link{summarize_deployments}()},
\code{\link{summarize_observations}()}
}
\concept{exploration functions}
