% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rai.R
\name{get_rai}
\alias{get_rai}
\title{Get Relative Abundance Index (RAI)}
\usage{
get_rai(x, species = "all", sex = NULL, life_stage = NULL)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{species}{Character with scientific names or common names (case
insensitive). If \code{"all"} (default) all scientific names are automatically
selected.}

\item{sex}{Character defining the sex class to filter on, e.g. \code{"female"} or
\code{c("male", "unknown")}. If \code{NULL} (default) all observations of all sex
classes are taken into account.}

\item{life_stage}{Character vector defining the life stage class to filter
on, e.g. \code{"adult"} or \code{c("subadult", "adult")}. If \code{NULL} (default) all
observations of all life stage classes are taken into account.}
}
\value{
A tibble data frame with the following columns: - \code{deploymentID}:
Deployment unique identifier. - \code{scientificName}: Scientific name. - \code{rai}:
Relative abundance index.
}
\description{
Gets the RAI (Relative Abundance Index) per deployment. The RAI is normalized
using 100 days deployment activity. In other words:
\code{RAI = 100 * (n/effort)} where \code{n} is the number of observations as
calculated via \code{get_n_obs()} and \code{effort} is the effort in days as calculated
via \code{get_effort()}.
}
\examples{
# Calculate RAI for all species
get_rai(x) # species = "all" by default, so equivalent of
get_rai(x, species = "all")

# Selected species
get_rai(x, species = c("Anas platyrhynchos", "Martes foina"))

# With vernacular names, even mixing languages
get_rai(x, species = c("mallard", "steenmarter"))

# Mixed scientific and vernacular names
get_rai(x, species = c("Anas platyrhynchos", "steenmarter"))

# Species parameter is case insensitive
get_rai(x, species = c("ANAS plAtyRhynChOS"))

# Specify sex
get_rai(x, sex = "female")
get_rai(x, sex = c("female", "unknown"))

# Specify life stage
get_rai(x, life_stage = "adult")
get_rai(x, life_stage = c("adult", "subadult"))
}
\seealso{
Other exploration functions: 
\code{\link{get_cam_op}()},
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_n_species}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_record_table}()},
\code{\link{get_scientific_name}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
