% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rai.R
\name{get_rai_individuals}
\alias{get_rai_individuals}
\title{Get Relative Abundance Index (RAI) based on number of individuals}
\usage{
get_rai_individuals(x, species = "all")
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{species}{Character with scientific names or common names (case
insensitive).
If \code{"all"} (default) all scientific names are automatically selected.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{scientificName}: Scientific name.
\item \code{rai}: Relative abundance index.
}
}
\description{
Function to get the RAI (Relative Abundance Index) per deployment based on
number of detected individuals instead of the number of observations. The RAI
is normalized using 100 days deployment activity. In other words:
\code{RAI = 100 * (n/effort)} where \code{n} is the number of individuals as calculated
via \code{get_n_individuals()} and \code{effort} is the effort in days as calculated
via \code{get_effort()}.
}
\examples{
x <- example_dataset()

# Calculate RAI based on number of individuals
get_rai_individuals(x) # species = "all" by default, so equivalent of
get_rai_individuals(x, species = "all")

# Selected species
get_rai_individuals(
  x,
  species = c("Anas platyrhynchos", "Martes foina")
)

# With common names, also mixing up languages
get_rai_individuals(x, species = c("mallard", "steenmarter"))

# Mixed scientific and vernacular names
get_rai_individuals(x, species = c("Anas platyrhynchos", "beech marten"))

# Species parameter is case insensitive
get_rai_individuals(x, species = c("ANAS plAtyRhynChOS"))

# Use `filter_observations()` to filter on life stage
x \%>\%
  filter_observations(lifeStage == "adult") \%>\%
  get_rai_individuals()

# Use `filter_observations()` to filter on sex
x \%>\%
  filter_observations(sex == "female") \%>\%
  get_rai_individuals()
}
\seealso{
Other exploration functions: 
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{n_species}()},
\code{\link{summarize_deployments}()}
}
\concept{exploration functions}
