% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_record_table.R
\name{get_record_table}
\alias{get_record_table}
\title{Get the record table}
\usage{
get_record_table(
  x,
  stationCol = "locationName",
  exclude = NULL,
  minDeltaTime = 0,
  deltaTimeComparedTo = NULL,
  removeDuplicateRecords = TRUE
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{stationCol}{Character name of the column containing stations.
Default: \code{"locationName"}.}

\item{exclude}{Character vector of scientific names to be excluded from the
record table. Default: \code{NULL}.}

\item{minDeltaTime}{Time difference between records of the same
species at the same station to be considered independent (in minutes).
Default: 0.}

\item{deltaTimeComparedTo}{One of \code{"lastIndependentRecord"} or
\code{"lastRecord"}.
For two records to be considered independent, the second one must be at
least \code{minDeltaTime} minutes after the last independent record of the same
species (\code{deltaTimeComparedTo = "lastIndependentRecord"}), or
\code{minDeltaTime} minutes after the last record (\code{deltaTimeComparedTo = "lastRecord"}).
If \code{minDeltaTime} is 0, \code{deltaTimeComparedTo} must be \code{NULL} (default).}

\item{removeDuplicateRecords}{Logical. If there are several records of the
same species, but e.g. different \code{sex} or \code{lifeStage}, at the same station
at exactly the same time, show only one? Default: \code{TRUE}. Duplicates are
removed by keeping only the first observation in the observation table.}
}
\value{
A tibble data frame containing species records and additional
information about stations, date, time and further metadata, such as
filenames and directories of the images (media) linked to the species
records.
Some more details about the columns returned:
\itemize{
\item \code{Station}: Character, station names, as found in the deployment column
defined in parameter \code{stationCol}.
\item \code{Species}: Character, the scientific name of the observed species.
\item \code{n}: Numeric, the number of observed individuals (renamed from
\href{https://camtrap-dp.tdwg.org/data/#observations.count}{\code{count}} in the
observations table).
\item \code{DateTimeOriginal}: Datetime object, as found in column \code{eventStart} of
\code{observations}, in UTC format.
\item \code{Date}: Date object, the date part of \code{DateTimeOriginal}, in UTC format.
\item \code{Time}: Character, the time part of \code{DateTimeOriginal} in UTC format.
\item \code{delta.time.secs}: Numeric, the duration in seconds from the previous
independent record of a given species at a certain location.
\item \code{delta.time.mins}: Numeric, the duration in minutes from the previous
independent record of a given species at a certain location.
\item \code{delta.time.hours}: Numeric, the duration in hours from the previous
independent record of a given species at a certain location.
\item \code{delta.time.days}: Numeric, the duration in days from the previous
independent record of a given species at a certain location.
\item \code{Directory}: List, file paths of the images linked to the given record,
as defined in column \code{filePath} of \code{media}.
\item \code{Filename}: List, file names of the images linked to the given record,
as defined in column \code{fileName} of \code{media}.
\item \code{Latitude}: Numeric, latitude of the station, based on \code{deploymentID}
of the observations.
\item \code{Longitude}: Numeric, longitude of the station, based on \code{deploymentID}
of the observations.
\item \code{clock}: Numeric, clock time in radians.
\item \code{solar}: Numeric, solar time in radians. Calculated using
\code{overlap::sunTime}, which essentially uses the approach described in
Nouvellet et al. (2012) \doi{10.1111/j.1469-7998.2011.00864.x}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Creates the record table from a Camera Trap Data Package and so tabulating
species records. Only event-based observations and their corresponding media
are taken into account. The record table is a concept developed within the
camtrapR package, see \href{https://jniedballa.github.io/camtrapR/articles/camtrapr3.html}{this article}. See also the
function documentation for \href{https://jniedballa.github.io/camtrapR/reference/recordTable.html}{camtrapR::recordTable()}.

This function is deprecated. Please use \code{\link[camtrapR:recordTable]{camtrapR::recordTable()}}. The
package camtrapR supports Camera Trap Data Packages, see the function
\code{\link[camtrapR:readCamtrapDP]{camtrapR::readCamtrapDP()}}.

\strong{Note}: All dates and times are expressed in UTC format.
}
\examples{
library(lubridate)

x <- example_dataset()
get_record_table(x)

# Create a new Camera Trap Data Package with dependent observations only for
# demonstration.
obs <- observations(x)
obs[obs$observationID == "9e191d10",]$scientificName <- "Martes foina"
x_dep <- x
observations(x_dep) <- obs

# Set a minDeltaTime of 100 minutes from last record
get_record_table(
  x_dep,
  minDeltaTime = 100,
  deltaTimeComparedTo = "lastRecord"
)

# Differences can occur between `deltaTimeCoparedTo` = `"lastRecord"` and
# `"lastIndependentRecord"`
obs <- observations(x)
obs[obs$eventID == "02ae9f43", "eventStart"] <- 
  as_datetime("2020-08-02 05:10:20")

med <- media(x) 
rows_to_update <- which(med$eventID == "02ae9f43") 
med[rows_to_update, "timestamp"] <- as_datetime("2020-08-02 05:10:20") 

x_modified <- x
observations(x_modified) <- obs
media(x_modified) <- med

rec_last_indep <- get_record_table(
  x_modified,
  minDeltaTime = 10,
  deltaTimeComparedTo = "lastIndependentRecord"
)

rec_last <- get_record_table(
  x_modified,
  minDeltaTime = 10,
  deltaTimeComparedTo = "lastRecord"
)

# Exclude observations of Anas platyrhynchos.
get_record_table(x, exclude = "Anas platyrhynchos")

# Specify column to pass station names
get_record_table(x, stationCol = "locationID")

# Include "duplicates", i.e. records of same species at same time, but
# different attributes, such as life stage or sex.
get_record_table(
 x,
 removeDuplicateRecords = FALSE
)
}
\seealso{
Other deprecated camtrapR-derived functions:
\code{\link[=get_cam_op]{get_cam_op()}},
\code{\link[=get_detection_history]{get_detection_history()}}
}
\concept{deprecated camtrapR-derived functions}
