% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_record_table.R
\name{get_record_table}
\alias{get_record_table}
\title{Get record table}
\usage{
get_record_table(
  datapkg,
  ...,
  stationCol = "location_name",
  exclude = NULL,
  minDeltaTime = 0,
  deltaTimeComparedTo = NULL
)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{...}{filter predicates for filtering on deployments}

\item{stationCol}{(character) name of the column containing stations.
Default: \code{"location_name"}}

\item{exclude}{(character) vector of species names (scientific names or
vernacular names) to be excluded from the record table. Default: \code{NULL}}

\item{minDeltaTime}{(integer) time difference between records of the same
species at the same station to be considered independent (in minutes).
Default: 0}

\item{deltaTimeComparedTo}{(character) \code{"lastIndependentRecord"} or
\code{"lastRecord"}. For two records to be considered independent, must the
second one be at least \code{minDeltaTime} minutes after the last independent
record of the same species (\code{deltaTimeComparedTo = "lastIndependentRecord"}
), or \code{minDeltaTime} minutes after the last record (
\code{deltaTimeComparedTo = "lastRecord"})? If \code{minDeltaTime} is 0,
\code{deltaTimeComparedTo} must be NULL (deafult)}
}
\value{
A (tibble) data frame containing species records and additional
information about stations, date, time and further metadata, such as
filenames and directories of the images (multimedia) linked to the species
records. Some more details about the columns returned:
\enumerate{
\item \code{Station}: character, station names, as found in the deployment column
defined  in argument \code{stationCol}
\item \code{Species}: character, the scientific name of the observed species
\item \code{DateTimeOriginal}: datetime object, as found in column \code{timestamp} of
\code{observations}
\item \code{Date}: date object, the date part of \code{DateTimeOriginal}
\item \code{Time}: character, the time part of \code{DateTimeOriginal}
\item \code{delta.time.secs}: numeric, the duration in seconds from the previous
independent record of a given species at a certain location
\item \code{delta.time.mins}: numeric, the duration in minutes from the previous
independent record of a given species at a certain location
\item \code{delta.time.hours}: numeric, the duration in hours from the previous
independent record of a given species at a certain location
\item \code{delta.time.days}: numeric, the duration in days from the previous
independent record of a given species at a certain location
\item \code{Directory}: list, file paths of the images linked to the given record,
as defined in column \code{file_path} of \code{multimedia}
\item \code{Filename}: list, file names of the images linked to the given record,
as defined in column \code{file_name} of \code{multimedia}
}
}
\description{
This function calculate the record table from a camera trap data package
and so tabulating species records. The record table is a concept develoepd
within the camtrapR package: see article
https://jniedballa.github.io/camtrapR/articles/camtrapr3.html. See also the
camtrapR's function documentation
\href{https://jniedballa.github.io/camtrapR/reference/recordTable.html}{recordTable}
}
\examples{
get_record_table(camtrapdp)

# set a minDeltaTime of 20 minutes from last independent record for filtering
# out not independent observations
get_record_table(camtrapdp,
    minDeltaTime = 20,
    deltaTimeComparedTo = "lastIndependentRecord")

# set a minDeltaTime of 20 minutes from last record for filtering out not
# independent observations
get_record_table(camtrapdp,
    minDeltaTime = 20,
    deltaTimeComparedTo = "lastRecord")

# exclude observations of Norway Rat
# exclude is case insensitive and vernacular names allowed
get_record_table(camtrapdp, exclude = "Norway raT")

# specify column to pass station names
get_record_table(camtrapdp,
    stationCol = "location_id",
    minDeltaTime = 20,
    deltaTimeComparedTo = "lastRecord")
# applying filter(s) on deployments, e.g. deployments with latitude >= 51.28
get_record_table(camtrapdp, pred_gte("latitude", 51.28))
}
