% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_record_table.R
\name{get_record_table}
\alias{get_record_table}
\title{Get record table}
\usage{
get_record_table(
  x,
  stationCol = "locationName",
  exclude = NULL,
  minDeltaTime = 0,
  deltaTimeComparedTo = NULL,
  removeDuplicateRecords = TRUE
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{stationCol}{Character name of the column containing stations.
Default: \code{"locationName"}.}

\item{exclude}{Character vector of species names (scientific names or
vernacular names) to be excluded from the record table.
Default: \code{NULL}.}

\item{minDeltaTime}{Time difference between records of the same
species at the same station to be considered independent (in minutes).
Default: 0.}

\item{deltaTimeComparedTo}{One of \code{"lastIndependentRecord"} or
\code{"lastRecord"}.
For two records to be considered independent, the second one must be at
least \code{minDeltaTime} minutes after the last independent record of the same
species (\code{deltaTimeComparedTo = "lastIndependentRecord"}), or
\code{minDeltaTime} minutes after the last record (\code{deltaTimeComparedTo = "lastRecord"}).
If \code{minDeltaTime} is 0, \code{deltaTimeComparedTo} must be \code{NULL} (default).}

\item{removeDuplicateRecords}{Logical.
If there are several records of the same species at the same station at
exactly the same time, show only one?}
}
\value{
A tibble data frame containing species records and additional
information about stations, date, time and further metadata, such as
filenames and directories of the images (media) linked to the species
records.
Some more details about the columns returned:
\itemize{
\item \code{Station}: Character, station names, as found in the deployment column
defined in parameter \code{stationCol}.
\item \code{Species}: Character, the scientific name of the observed species.
\item \code{DateTimeOriginal}: Datetime object, as found in column \code{timestamp} of
\code{observations}, in UTC format.
\item \code{Date}: Date object, the date part of \code{DateTimeOriginal}, in UTC format.
\item \code{Time}: Character, the time part of \code{DateTimeOriginal} in UTC format.
\item \code{delta.time.secs}: Numeric, the duration in seconds from the previous
independent record of a given species at a certain location.
\item \code{delta.time.mins}: Numeric, the duration in minutes from the previous
independent record of a given species at a certain location.
\item \code{delta.time.hours}: Numeric, the duration in hours from the previous
independent record of a given species at a certain location.
\item \code{delta.time.days}: Numeric, the duration in days from the previous
independent record of a given species at a certain location.
\item \code{Directory}: List, file paths of the images linked to the given record,
as defined in column \code{filePath} of \code{media}.
\item \code{Filename}: List, file names of the images linked to the given record,
as defined in column \code{fileName} of \code{media}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=camtrapR_recordTable]{camtrapR_recordTable()}} instead.
}
\examples{
x <- example_dataset()
camtrapR_recordTable(x)

# Set a minDeltaTime of 20 minutes from last independent record for filtering
# out not independent observations
x_dependent <- x
x_dependent$data$observations[4,"timestamp"] <- lubridate::as_datetime("2020-07-29 05:55:00")
camtrapR_recordTable(
  x_dependent,
  minDeltaTime = 20,
  deltaTimeComparedTo = "lastIndependentRecord"
)

# Set a minDeltaTime of 20 minutes from last record for filtering out not
# independent observations
camtrapR_recordTable(
  x_dependent,
  minDeltaTime = 20,
  deltaTimeComparedTo = "lastRecord"
)

# Exclude observations of mallard
# Exclude is case insensitive and vernacular names are allowed
camtrapR_recordTable(x, exclude = "wilde eend")

# Specify column to pass station names
camtrapR_recordTable(
  x,
  stationCol = "locationID",
  minDeltaTime = 20,
  deltaTimeComparedTo = "lastRecord"
)

# How to deal with duplicates
x_dup <- x
# create a duplicate at 2020-07-29 05:46:48, location: B_DL_val 5_beek kleine vijver
x_dup$data$observations[4,"sequenceID"] <- purrr::pluck(
  observations(x_dup),
  "sequenceID",
  3
)
x_dup$data$observations[4, "deploymentID"] <- purrr::pluck(
  observations(x_dup),
  "deploymentID",
  3
)
x_dup$data$observations[4, "timestamp"] <- purrr::pluck(
  observations(x_dup),
  "timestamp",
  3
)

# duplicates are removed by default by camtrapR_recordTable()
camtrapR_recordTable(x_dup)

# duplicate not removed
camtrapR_recordTable(x_dup, removeDuplicateRecords = FALSE)
}
