% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scientific_name.R
\name{get_scientific_name}
\alias{get_scientific_name}
\title{Get scientific name based on its vernacular name}
\usage{
get_scientific_name(datapkg, vernacular_name)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}, i.e. a list containing three data.frames:
\enumerate{
\item \code{observations} 2. \code{deployments} 3. \code{multimedia}
}

and a list with metadata: \code{datapackage}}

\item{vernacular_name}{a character vector with the vernacular name(s) of the
species as found in the \code{observations} table of the camera trap data
package}
}
\value{
a character vector of scientific name(s)
}
\description{
This function returns the scientific name(s) of a vector of vernacular names
based on the observations table of the given camera trap data package.
}
\examples{
get_scientific_name(camtrapdp, "Norway Rat")

}
