% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scientific_name.R
\name{get_scientific_name}
\alias{get_scientific_name}
\title{Get scientific name for vernacular name}
\usage{
get_scientific_name(
  package = NULL,
  vernacular_name,
  datapkg = lifecycle::deprecated()
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{vernacular_name}{Character vector with input vernacular name(s).}

\item{datapkg}{Deprecated.
Use \code{package} instead.}
}
\value{
Character vector of scientific name(s).
}
\description{
Gets the scientific name for one or more vernacular names based on the
taxonomic information found in the metadata (\code{package$taxonomic}).
The match is performed case insensitively.
If a vernacular name is not valid, an error is returned
}
\examples{
# One or more vernacular names
get_scientific_name(mica, "beech marten")
get_scientific_name(mica, c("beech marten", "mallard"))

# Vernacular names can be passed in different languages
get_scientific_name(mica, c("beech marten", "wilde eend"))

# Search is performed case insensitively
get_scientific_name(mica, c("MaLLarD"))

\dontrun{
# An error is returned if at least one invalid vernacular name is passed
get_scientfic_name(mica, "this is a bad vernacular name")

# A scientific name is an invalid vernacular name of course
get_scientific_name(mica, c("Castor fiber", "wilde eend"))
}
}
\seealso{
Other exploration functions: 
\code{\link{get_cam_op}()},
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_n_species}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_record_table}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
