% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scientific_name.R
\name{get_scientific_name}
\alias{get_scientific_name}
\title{Get scientific name for vernacular name}
\usage{
get_scientific_name(x, vernacular_name)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{vernacular_name}{Character vector with input vernacular name(s).}
}
\value{
Character vector of scientific name(s).
}
\description{
Gets the scientific name for one or more vernacular names based on the
taxonomic information found in the metadata (\code{x$taxonomic}).
The match is performed case insensitively.
If a vernacular name is not valid, an error is returned
}
\examples{
x <- example_dataset()

# One or more vernacular names
get_scientific_name(x, "beech marten")
get_scientific_name(x, c("beech marten", "mallard"))

# Vernacular names can be passed in different languages
get_scientific_name(x, c("beech marten", "wilde eend"))

# Search is performed case insensitively
get_scientific_name(x, c("MaLLarD"))

\dontrun{
# An error is returned if at least one invalid vernacular name is passed
get_scientific_name(x, "this is a bad vernacular name")

# A scientific name is an invalid vernacular name of course
get_scientific_name(x, c("Castor fiber", "wilde eend"))
}
}
\seealso{
Other exploration functions: 
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_species}()},
\code{\link{n_species}()}
}
\concept{exploration functions}
