% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get species}
\usage{
get_species(package = NULL, datapkg = lifecycle::deprecated())
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{datapkg}{Deprecated. Use \code{package} instead.}
}
\value{
Tibble data.frame with all scientific names and vernacular names of
the identified species.
}
\description{
Function to get all identified species
}
\examples{
get_species(mica)
}
