% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{locations}
\alias{locations}
\title{Get or set locations}
\usage{
locations(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}
}
\description{
Re-export of \code{\link[camtrapdp:locations]{camtrapdp::locations()}}. See that function for full
documentation.
}
\seealso{
Other accessor functions: 
\code{\link{add_coordinates}()},
\code{\link{deployments}()},
\code{\link{deployments<-}()},
\code{\link{events}()},
\code{\link{get_species}()},
\code{\link{media}()},
\code{\link{media<-}()},
\code{\link{observations}()},
\code{\link{observations<-}()},
\code{\link{taxa}()}
}
\concept{accessor functions}
