% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dep.R
\name{map_dep}
\alias{map_dep}
\title{Visualize deployments features}
\usage{
map_dep(
  deployments,
  observations,
  feature,
  cluster = TRUE,
  hover_columns = c("n", "deployment_id", "location_id", "location_name", "latitude",
    "longitude", "start", "end"),
  relative_color_scale = TRUE,
  max_color_scale = NULL,
  radius_range = c(10, 50)
)
}
\arguments{
\item{deployments}{a tibble (data.frame) containing deployments}

\item{observations}{a tibble (data.frame) containing observations}

\item{feature}{character, one of: - \code{n_species} - \code{n_obs}}

\item{cluster}{a logical value indicating whether using the cluster option
while visualizing maps. Default: TRUE}

\item{hover_columns}{character with the name of the columns to use for
showing location deployment information while hovering the mouse over. One
or more from deployment columns. Use \code{NULL} to disable hovering. Default
information:
\itemize{
\item \code{n}: number of species/observations (column created internally by a
\verb{get_n_*()} function)
\item \code{start}: start deployment
\item \code{end}: end deployment
\item \code{deployment_id}
\item \code{location_id}
\item \code{location_name}
\item \code{latitude}
\item \code{longitude}
}

See \href{https://tdwg.github.io/camtrap-dp/data/#deployments}{section Deployment of Camtrap DP standard} for the full
list of all columns you can use}

\item{relative_color_scale}{a logical indicating whether to use a relative
color scale (\code{TRUE}) or an absolute scale (\code{FALSE}). If absolute scale is
used, specify a valid \code{max_color_scale}}

\item{max_color_scale}{a number indicating the max value used in the color
scale}

\item{radius_range}{a vector of length 2 containing the lower and upper limit
of the circle radius. The lower value is used for deployments with zero
observations/identified species. The upper value for the deployment(s) with
the highest number of identified species/observations. Default: \code{c(10, 50)}}
}
\value{
a leaflet map
}
\description{
This function visualizes deployments features such as number of species
detected or number of observations on a dynamic map. The circle size and
color are proportional to the mapped feature. Deployments without
observations are shown as gray circles and a message is returned.
}
\details{
Possible filtering about time period, ... (to be added)
}
\examples{
\dontrun{
# show number of species
map_dep(
  camtrapdp$deployments,
  camtrapdp$observations,
  "n_species"
)

# show number of observations
map_dep(
  camtrapdp$deployments,
  camtrapdp$observations,
  "n_obs"
)

# cluster disabled
map_dep(camtrapdp$deployments,
  camtrapdp$observations,
  "n_species"
  cluster = FALSE
)

# show only number of observations and location name while hovering
map_dep(camtrapdp$deployments,
  camtrapdp$observations,
  "n_obs",
  hover_columns = c("location_name", "n")
)

# use absolute scale for colors
map_dep(camtrapdp$deployments,
  camtrapdp$observations,
  "n_species",
  relative_color_scale = FALSE,
  max_color_scale = 4
)

# change max and min size circles
map_dep(
  camtrapdp$deployments,
  camtrapdp$observations,
  "n_obs",
  radius_range = c(40, 150)
)
}
}
