% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dep.R
\name{map_dep}
\alias{map_dep}
\title{Visualize deployments features}
\usage{
map_dep(
  package = NULL,
  feature,
  ...,
  species = NULL,
  sex = NULL,
  life_stage = NULL,
  effort_unit = NULL,
  cluster = TRUE,
  hover_columns = c("n", "species", "deploymentID", "locationID", "locationName",
    "latitude", "longitude", "start", "end"),
  palette = "inferno",
  zero_values_show = TRUE,
  zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/000000/multiply.png",
  zero_values_icon_size = 10,
  relative_scale = TRUE,
  max_scale = NULL,
  radius_range = c(10, 50),
  datapkg = lifecycle::deprecated()
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{feature}{Deployment feature to visualize. One of:
\itemize{
\item \code{"n_species"}: number of identified species
\item \code{"n_obs"}: number of observations
\item \code{"n_individuals"}: number of individuals
\item \code{"rai"}: Relative Abundance Index
\item \code{"effort"}: effort (duration) of the deployment
}}

\item{...}{Filter predicates for subsetting deployments.}

\item{species}{Character with a scientific name. Required for  \code{rai},
optional for \code{n_obs}. Default: \code{NULL}.}

\item{sex}{Character defining the sex class to filter on, e.g. \code{"female"}.
If \code{NULL}, default, all observations of all sex classes are taken into
account. Optional argument for \code{n_obs} and \code{n_individuals}.}

\item{life_stage}{Character vector defining the life stage class to filter on, e.g.
\code{"adult"} or \code{c("subadult", "adult")}. If \code{NULL}, default, all observations
of all life stage classes are taken into account. Optional argument for \code{n_obs}
and \code{n_individuals}.}

\item{effort_unit}{Time unit to use while visualizing deployment effort
(duration). One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
}}

\item{cluster}{Logical value indicating whether using the cluster option
while visualizing maps. Default: TRUE.}

\item{hover_columns}{Character vector with the name of the columns to use for
showing location deployment information while hovering the mouse over. One
or more from deployment columns. Use \code{NULL} to disable hovering. Default
information:
\itemize{
\item \code{n}: number of species, number of observations, RAI or effort (column
created internally by a \verb{get_*()} function)
\item \code{species}: species name(s)
\item \code{start}: start deployment
\item \code{end}: end deployment -
\item \code{deploymentID} deployment unique identifier
\item \code{locationID} location unique identifier
\item \code{locationName} location name
\item \code{latitude}
\item \code{longitude}
}

See \href{https://tdwg.github.io/camtrap-dp/data/#deployments}{section Deployment of Camtrap DP standard} for the full
list of columns you can use.}

\item{palette}{The palette name or the color function that values will be
mapped to.
Typically one of the following:
\itemize{
\item A character vector of RGB or named colors. Examples: \verb{c("#000000",   "#0000FF", "#FFFFFF"))},\verb{topo.colors(10))}.
\item the full name of a RColorBrewer palette, e.g. "BuPu" or "Greens", or
viridis palette: \code{"viridis"}, \code{"magma"}, \code{"inferno"} or \code{"plasma"}
For more options, see argument \code{palette} of \code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}}.
}}

\item{zero_values_show}{Logical indicating whether to show deployments with
zero values. Default: \code{TRUE}.}

\item{zero_values_icon_url}{character with URL to icon for showing
deployments with zero values. Default: a cross (multiply symbol)
\code{"https://img.icons8.com/ios-glyphs/30/000000/multiply.png"}.}

\item{zero_values_icon_size}{a number to set the size of the icon to show
deployments with wero values. Default: 10.}

\item{relative_scale}{Logical indicating whether to use a relative color
and radius scale (\code{TRUE}) or an absolute scale (\code{FALSE}). If absolute scale
is used, specify a valid \code{max_scale}.}

\item{max_scale}{Number indicating the max value used to map color
and radius.}

\item{radius_range}{Vector of length 2 containing the lower and upper limit
of the circle radius. The lower value is used for deployments with zero
feature value, i.e. no observations, no identified species, zero RAI or
zero effort. The upper value is used for the deployment(s) with the highest
feature value (\code{relative_scale} = \code{TRUE}) or \code{max_scale} (\code{relative_scale}
= \code{FALSE}). Default: \code{c(10, 50)}.}

\item{datapkg}{Deprecated. Use \code{package} instead.}
}
\value{
Leaflet map.
}
\description{
This function visualizes deployments features such as number of detected
species, number of observations and RAI on a dynamic map. The circle size and
color are proportional to the mapped feature. Deployments without
observations are shown as gray circles and a message is returned.
}
\examples{
\dontrun{
# show number of species
map_dep(
  mica,
  "n_species"
)

# show number of observations  (observations of unidentified species included
if any)
map_dep(
  mica,
  "n_obs"
)

# show number of observations of Anas platyrhynchos
map_dep(
  mica,
  "n_obs",
  species = "Anas platyrhynchos"
)

# show number of observations of subadult individuals of Anas strepera
map_dep(
  mica,
  "n_obs",
  species = "Anas strepera",
  life_stage = "subadult"
)

# show number of observations of female or unknown individuals of gadwall
map_dep(
  mica,
  "n_obs",
  species = "gadwall",
  sex = c("female", "unknown")
)

# show number of individuals (individuals of unidentified species included if
any)
map_dep(
  mica,
  "n_individuals"
)

# same filters by life stage and sex as for number of observations apply
map_dep(
  mica,
  "n_individuals",
  species = "Anas strepera",
  sex = "female",
  life_stage = "adult"
)

# show RAI
map_dep(
  mica,
  "rai",
  species = "Anas strepera"
)

# same filters by life_stage and sex as for number of observations apply
map_dep(
  mica,
  "rai",
  species = "Anas strepera",
  sex = "female",
  life_stage = "adult"
)

# show RAI calculated by using number of detected individuals
map_dep(
  mica,
  "rai_individuals",
  species = "Anas strepera"
)

# same filters by life stage and sex as for basic RAI apply
map_dep(
  mica,
  "rai_individuals",
  species = "Anas strepera",
  sex = "female",
  life_stage = "adult"
)

# show effort (days)
map_dep(
  mica,
  "effort",
  effort_unit = "day"
)

# show effort (months)
map_dep(
  mica,
  "effort",
  effort_unit = "month"
)

# use viridis palette (viridis palettes)
map_dep(
  mica,
  "n_obs",
  palette = "viridis"
)

# use "BuPu" color palette (RColorBrewer palettes)
map_dep(
  mica,
  "n_obs",
  palette = "BuPu"
)

# use a palette defined by color names
map_dep(
  mica,
  "n_obs",
  palette = c("black", "blue", "white")
)

#' # use a palette defined by hex colors
map_dep(
  mica,
  "n_obs",
  palette = c("#000000", "#0000FF", "#FFFFFF")
)

#' # do not show deployments with zero values
map_dep(
  mica,
  "n_obs",
  life_stage = "subadult",
  zero_values_show = FALSE
)

#' # use same icon but but a non default color for zero values deployments,
e.g. red (hex: E74C3C)
map_dep(
  mica,
  "n_obs",
  life_stage = "subadult",
  zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/E74C3C/multiply.png"
)

# or yellow (F1C40F)
map_dep(
  mica,
  "n_obs",
  life_stage = "subadult",
  zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/F1C40F/multiply.png"
)

# use another icon via a different URL, e.g. the character Fry from Futurama
in green (2ECC71)
map_dep(
  mica,
  "n_obs",
  life_stage = "subadult",
  zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/2ECC71/futurama-fry.png"
)

# set size of the icon for zero values deployments
map_dep(
  mica,
  "n_obs",
  life_stage = "subadult",
  zero_values_icon_size = 30
)

# disable cluster
map_dep(
  mica,
  "n_species",
  cluster = FALSE
)

# show only number of observations and location name while hovering
map_dep(
  mica,
  "n_obs",
  hover_columns = c("locationName", "n")
)

# use absolute scale for colors and radius
map_dep(mica,
  "n_species",
  relative_scale = FALSE,
  max_scale = 4
)

# change max and min size circles
map_dep(
  mica,
  "n_obs",
  radius_range = c(40, 150)
)
}
}
\seealso{
Check documentation about filter predicates: \code{\link[=pred]{pred()}}, \code{\link[=pred_in]{pred_in()}},
\code{\link[=pred_and]{pred_and()}}, ...
}
