% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dep.R
\name{map_dep}
\alias{map_dep}
\title{Visualize deployments features}
\usage{
map_dep(
  datapkg,
  feature,
  ...,
  species = NULL,
  sex = NULL,
  age = NULL,
  effort_unit = NULL,
  cluster = TRUE,
  hover_columns = c("n", "species", "deployment_id", "location_id", "location_name",
    "latitude", "longitude", "start", "end"),
  relative_scale = TRUE,
  max_scale = NULL,
  radius_range = c(10, 50)
)
}
\arguments{
\item{datapkg}{a camera trap data package object, as returned by
\code{read_camtrap_dp()}.}

\item{feature}{character, deployment feature to visualize. One of:
\itemize{
\item \code{n_species}: number of identified species
\item \code{n_obs}: number of observations
\item \code{n_individuals}: number of individuals
\item \code{rai}: Relative Abundance Index
\item \code{effort}: effort (duration) of the deployment
}}

\item{...}{filter predicates for subsetting deployments}

\item{species}{a character with a scientific name. Required for  \code{rai},
optional for \code{n_obs}. Default: \code{NULL}}

\item{sex}{a character defining the sex class to filter on, e.g. \code{"female"}.
If \code{NULL}, default, all observations of all sex classes are taken into
account. Optional argument for \code{n_obs} and \code{n_individuals}}

\item{age}{a character vector defining the age class to filter on, e.g.
\code{"adult"} or \code{c("subadult", "adult")}. If \code{NULL}, default, all observations
of all age classes are taken into account. Optional argument for \code{n_obs}
and \code{n_individuals}}

\item{effort_unit}{time unit to use while visualizing deployment effort
(duration). One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
\item \code{NULL} (default) duration objects (e.g. 2594308s (~4.29 weeks)) are shown
while hovering and seconds shown in legend
}}

\item{cluster}{a logical value
indicating whether using the cluster option while visualizing maps.
Default: TRUE}

\item{hover_columns}{character with the name of the columns to use for
showing location deployment information while hovering the mouse over. One
or more from deployment columns. Use \code{NULL} to disable hovering. Default
information:
\itemize{
\item \code{n}: number of species, number of observations, RAI or effort (column
created internally by a \verb{get_*()} function)
\item \code{species}: species name(s)
\item \code{start}: start deployment
\item \code{end}: end deployment -
\item \code{deployment_id} deployment unique identifier
\item \code{location_id} location unique identifier
\item \code{location_name} location name
\item \code{latitude}:
\item \code{longitude}:
}

See \href{https://tdwg.github.io/camtrap-dp/data/#deployments}{section Deployment of Camtrap DP standard} for the full
list of columns you can use}

\item{relative_scale}{a logical indicating whether to use a relative color
and radius scale (\code{TRUE}) or an absolute scale (\code{FALSE}). If absolute scale
is used, specify a valid \code{max_scale}}

\item{max_scale}{a number indicating the max value used to map color
and radius}

\item{radius_range}{a vector of length 2 containing the lower and upper limit
of the circle radius. The lower value is used for deployments with zero
feature value, i.e. no observations, no identified species, zero RAI or
zero effort. The upper value is used for the deployment(s) with the highest
feature value (\code{relative_scale} = \code{TRUE}) or \code{max_scale} (\code{relative_scale}
= \code{FALSE}). Default: \code{c(10, 50)}}
}
\value{
a leaflet map
}
\description{
This function visualizes deployments features such as number of detected
species, number of observations and RAI on a dynamic map. The circle size and
color are proportional to the mapped feature. Deployments without
observations are shown as gray circles and a message is returned.
}
\examples{
\dontrun{
# show number of species
map_dep(
  camtrapdp,
  "n_species"
)

# show number of observations  (observations of unidentified species included
if any)
map_dep(
  camtrapdp,
  "n_obs"
)

# show number of observations of Rattus norvegicus
map_dep(
  camtrapdp,
  "n_obs",
  species = "Rattus norvegicus"
)

# show number of observations of juvenile individuals of Anas platyrhynchos
map_dep(
  camtrapdp,
  "n_obs",
  species = "Anas platyrhynchos",
  age = "juvenile"
)

# show number of observations of female or undefined individuals of Anas platyrhynchos
map_dep(
  camtrapdp,
  "n_obs",
  species = "Anas platyrhynchos",
  sex = c("female", "undefined")
)

# show number of individuals (individuals of unidentified species included if
any)
map_dep(
  camtrapdp,
  "n_individuals"
)

# same filters by age and sex as for number of observations apply
map_dep(
  camtrapdp,
  "n_individuals",
  species = "Anas platyrhynchos",
  sex = "female",
  age = "adult"
)

# show RAI
map_dep(
  camtrapdp,
  "rai",
  species = "Rattus norvegicus"
)

# same filters by age and sex as for number of observations apply
map_dep(
  camtrapdp,
  "rai",
  species = "Anas platyrhynchos",
  sex = "female",
  age = "adult"
)

# show RAI calculated by using number of detected individuals
map_dep(
  camtrapdp,
  "rai_individuals",
  species = "Rattus norvegicus"
)

# same filters by age and sex as for basic RAI apply
map_dep(
  camtrapdp,
  "rai_individuals",
  species = "Anas platyrhynchos",
  sex = "female",
  age = "adult"
)

# show effort (basic duration in seconds)
map_dep(
  camtrapdp,
  "effort"
)

# show effort (days)
map_dep(
  camtrapdp,
  "effort",
  effort_unit = "day"
)

# show effort (months)
map_dep(
  camtrapdp,
  "effort",
  effort_unit = "month"
)

# cluster disabled
map_dep(camtrapdp,
  "n_species"
  cluster = FALSE
)

# show only number of observations and location name while hovering
map_dep(camtrapdp,
  "n_obs",
  hover_columns = c("location_name", "n")
)

# use absolute scale for colors and radius
map_dep(camtrapdp,
  "n_species",
  relative_scale = FALSE,
  max_scale = 4
)

# change max and min size circles
map_dep(
  camtrapdp,
  "n_obs",
  radius_range = c(40, 150)
)
}
}
\seealso{
Check documentation about filter predicates: \code{\link[=pred]{pred()}}, \code{\link[=pred_in]{pred_in()}}, \code{\link[=pred_and]{pred_and()}}, ...
}
