% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dep.R
\name{map_dep}
\alias{map_dep}
\title{Visualize deployments features}
\usage{
map_dep(
  x,
  feature,
  species = NULL,
  sex = NULL,
  life_stage = NULL,
  effort_unit = NULL,
  cluster = TRUE,
  hover_columns = c("n", "species", "deploymentID", "locationID", "locationName",
    "latitude", "longitude", "start", "end"),
  palette = "inferno",
  zero_values_show = TRUE,
  zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/000000/multiply.png",
  zero_values_icon_size = 10,
  na_values_show = TRUE,
  na_values_icon_url = "https://img.icons8.com/ios-glyphs/30/FA5252/multiply.png",
  na_values_icon_size = 10,
  relative_scale = TRUE,
  max_scale = NULL,
  radius_range = c(10, 50)
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{feature}{Deployment feature to visualize.
One of:
\itemize{
\item \code{n_species}: Number of identified species.
\item \code{n_obs}: Number of observations.
\item \code{n_individuals}: Number of individuals.
\item \code{rai}: Relative Abundance Index.
\item \code{rai_individuals}: Relative Abundance Index based on number of individuals.
\item \code{effort}: Effort (duration) of the deployment.
}}

\item{species}{Character with a scientific name.
Required for \code{rai}, optional for \code{n_obs}.
Default: \code{NULL}.}

\item{effort_unit}{Time unit to use while visualizing deployment effort
(duration).
One of:
\itemize{
\item \code{second}
\item \code{minute}
\item \code{hour}
\item \code{day}
\item \code{month}
\item \code{year}
If  \code{NULL} (default), the effort is returned in hours.
}}

\item{cluster}{Logical value indicating whether using the cluster option
while visualizing maps.
Default: \code{TRUE}.}

\item{hover_columns}{Character vector with the name of the columns to use for
showing location deployment information on mouse hover.
One or more from deployment columns.
Use \code{NULL} to disable hovering.
Default information:
\itemize{
\item \code{n}: Number of species, number of observations, RAI or effort (column
created internally by a \verb{get_*()} function).
\item \code{species}: Species name(s).
\item \code{start}: Start deployment.
\item \code{end}: End deployment.
\item \code{deploymentID}: Deployment unique identifier.
\item \code{locationID}: Location unique identifier.
\item \code{locationName}: Location name.
\item \code{latitude}
\item \code{longitude}
}

See the \href{https://camtrap-dp.tdwg.org/data/#deployments}{Deployment}
section of Camtrap DP for the full list of columns you can use.}

\item{palette}{The palette name or the colour function that values will be
mapped to.
Typically one of the following:
\itemize{
\item A character vector of RGB or named colours. Examples: \verb{c("#000000",   "#0000FF", "#FFFFFF"))},\verb{topo.colors(10))}.
\item The full name of a RColorBrewer palette, e.g. "BuPu" or "Greens", or
viridis palette: \code{"viridis"}, \code{"magma"}, \code{"inferno"} or \code{"plasma"}.
For more options, see parameter \code{palette} of \code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}}.
}}

\item{zero_values_show}{Logical indicating whether to show deployments with
zero values.
Default: \code{TRUE}.}

\item{zero_values_icon_url}{Character with URL to icon for showing
deployments with zero values.
Default: a cross (multiply symbol)
\code{"https://img.icons8.com/ios-glyphs/30/000000/multiply.png"}.}

\item{zero_values_icon_size}{A number to set the size of the icon to show
deployments with zero values.
Default: 10.}

\item{na_values_show}{Logical indicating whether to show deployments with
zero values. Notice that only feature \code{"n_species"}
generates NA values.
Default: \code{TRUE}.}

\item{na_values_icon_url}{Character with URL to icon for showing
deployments with \code{NA} values. Notice that only feature \code{"n_species"}
generates NA values.
Default: a red cross (multiply symbol)
\code{"https://img.icons8.com/ios-glyphs/30/FA5252/multiply.png"}.}

\item{na_values_icon_size}{A number to set the size of the icon to show
deployments with \code{NA} values. Notice that only feature \code{"n_species"}
generates NA values.
Default: 10.}

\item{relative_scale}{Logical indicating whether to use a relative colour
and radius scale (\code{TRUE}) or an absolute scale (\code{FALSE}).
If absolute scale is used, specify a valid \code{max_scale}.}

\item{max_scale}{Number indicating the max value used to map colour
and radius.}

\item{radius_range}{Vector of length 2 containing the lower and upper limit
of the circle radius.
The lower value is used for deployments with zero feature value, i.e. no
observations, no identified species, zero RAI or zero effort.
The upper value is used for the deployment(s) with the highest feature
value (\code{relative_scale} = \code{TRUE}) or \code{max_scale} (\code{relative_scale}
= \code{FALSE}).
Default: \code{c(10, 50)}.}
}
\value{
Leaflet map.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=map_deployments]{map_deployments()}} instead.
}
\examples{
\dontrun{
x <- example_dataset()

# Show number of species
map_deployments(
  x,
  "n_species"
)

# Show number of observations (observations of unidentified species included
# if any)
map_deployments(
  x,
  "n_obs"
)

# Show number of observations of Anas platyrhynchos
map_deployments(
  x,
  "n_obs",
  species = "Anas platyrhynchos"
)

# Show number of individuals (individuals of unidentified species included if
# any)
map_deployments(
  x,
  "n_individuals"
)

# Show RAI
map_deployments(
  x,
  "rai",
  species = "Anas strepera"
)

# Show RAI calculated by using number of detected individuals
map_deployments(
  x,
  "rai_individuals",
  species = "Anas strepera"
)

# Show effort (hours)
map_deployments(
  x,
  "effort"
)
# Show effort (days)
map_deployments(
  x,
  "effort",
  effort_unit = "day"
)

# Use viridis palette (viridis palettes)
map_deployments(
  x,
  "n_obs",
  palette = "viridis"
)

# Use "BuPu" colour palette (RColorBrewer palettes)
map_deployments(
  x,
  "n_obs",
  palette = "BuPu"
)

# Use a palette defined by colour names
map_deployments(
  x,
  "n_obs",
  palette = c("black", "blue", "white")
)

# Use a palette defined by hex colours
map_deployments(
  x,
  "n_obs",
  palette = c("#000000", "#0000FF", "#FFFFFF")
)

# Do not show deployments with zero values
x \%>\%
  filter_observations(lifeStage == "subadult") \%>\%
  map_deployments(
    "n_obs",
    zero_values_show = FALSE
  )

# Use same icon but but a non default colour for zero values deployments,
# E.g. red (hex: E74C3C)
x \%>\%
  filter_observations(lifeStage == "subadult") \%>\%
  map_deployments(
    "n_obs",
    zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/E74C3C/multiply.png"
  )

# ... or yellow (F1C40F)
x \%>\%
  filter_observations(lifeStage == "subadult") \%>\%
  map_deployments(
    "n_obs",
    zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/F1C40F/multiply.png"
  )

# Use another icon via a different URL, e.g. the character Fry from Futurama
# in green (2ECC71)
x \%>\%
  filter_observations(lifeStage == "subadult") \%>\%
  map_deployments(
    "n_obs",
    zero_values_icon_url = "https://img.icons8.com/ios-glyphs/30/2ECC71/futurama-fry.png"
  )
  
# Set size of the icon for zero values deployments
x \%>\% 
  filter_observations(lifeStage == "subadult") \%>\%
  map_deployments("n_obs", zero_values_icon_size = 30)

# Use another icon url/size for visualizing NA values (`"n_species"` feature)
x \%>\%
  filter_observations(deploymentID != "00a2c20d") \%>\%
  map_deployments(
    feature = "n_species",
    na_values_icon_url = "https://img.icons8.com/ios-glyphs/30/E74C3C/futurama-fry.png",
    na_values_icon_size = 60
)

# Disable cluster
map_deployments(
  x,
  "n_species",
  cluster = FALSE
)

# Show only number of observations and location name while hovering
map_deployments(
  x,
  "n_obs",
  hover_columns = c("locationName", "n")
)

# Use absolute scale for colours and radius
map_deployments(
  x,
  "n_species",
  relative_scale = FALSE,
  max_scale = 4
)

# Change max and min size circles
map_deployments(
  x,
  "n_obs",
  radius_range = c(40, 150)
)
}
}
