% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{media}
\alias{media}
\alias{media<-}
\title{Get or set media}
\usage{
media(x)

media(x) <- value
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}

\item{value}{A data frame to assign as media.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with media.
}
\description{
Re-export of \code{\link[camtrapdp:media]{camtrapdp::media()}}.
}
\examples{
x <- example_dataset()
# Get media
media(x)

# Set media (not recommended outside a function)
media(x) <- head(media(x), 1)
}
\seealso{
Other accessor functions:
\code{\link[=contributors]{contributors()}},
\code{\link[=deployments]{deployments()}},
\code{\link[=events]{events()}},
\code{\link[=individuals]{individuals()}},
\code{\link[=locations]{locations()}},
\code{\link[=observations]{observations()}},
\code{\link[=taxa]{taxa()}}
}
\concept{accessor functions}
