% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\title{Retrieve media data from a Camera Trap Data Package}
\usage{
media(package)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}
}
\value{
A tibble with the deployments of the \code{package}
}
\description{
Retrieve media data from a Camera Trap Data Package
}
\examples{
media(mica)
}
