% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mica}
\alias{mica}
\title{Sample of camtrap-dp formatted data}
\format{
An object of class \code{list} of length 4.
}
\source{
https://github.com/tdwg/camtrap-dp/tree/ad0278ef86ef518dacfb306c598dce97667cfb81/example
}
\usage{
mica
}
\description{
A sample camera trap data package, based on the example included with
Camtrap DP, based on data from the MICA project. Also stored in \code{inst}.
}
\keyword{datasets}
