% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mica}
\alias{mica}
\title{Sample of Camtrap DP formatted data}
\format{
An object of class \code{datapackage} (inherits from \code{list}) of length 16.
}
\source{
\url{https://github.com/tdwg/camtrap-dp/tree/ad0278ef86ef518dacfb306c598dce97667cfb81/example}
}
\usage{
mica
}
\description{
A sample \href{https://tdwg.github.io/camtrap-dp}{Camera Trap Data Package} as
read by \code{\link[=read_camtrap_dp]{read_camtrap_dp()}}.
This sample is an extract from a larger Camera Trap Data Package saved as raw
data in \code{inst/extdata/mica_zenodo_5590881}.
To save memory, the \code{media.csv} resource of the latter Data Package is not
present.
The entire dataset can be downloaded from \href{https://zenodo.org/record/5590881}{Zenodo}.
}
\examples{
\dontrun{
# mica.rda was created with the code below.
mica <- read_camtrap_dp(
  system.file(
    "extdata/mica",
    "datapackage.json",
    package = "camtraptor"
  )
)
save(mica, file = "data/mica.rda")
}
}
\seealso{
Other sample data: 
\code{\link{animal_positions}},
\code{\link{dep_calib_models}}
}
\concept{sample data}
\keyword{datasets}
