% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mica}
\alias{mica}
\title{Sample of camtrap-dp formatted data}
\format{
An object of class \code{datapackage} (inherits from \code{list}) of length 16.
}
\source{
https://github.com/tdwg/camtrap-dp/tree/ad0278ef86ef518dacfb306c598dce97667cfb81/example
}
\usage{
mica
}
\description{
A sample camera trap data package, based on data from the MICA project. Also
stored in \code{inst}.
}
\details{
This sample is an extract from a bigger camera trap data package saved as raw
data in \code{inst/extdata/mica_zenodo_5590881}. For memory reasons \code{media.csv}
resource is not present. The entire data package can be downloaded from
\href{https://zenodo.org/record/5590881}{zenodo}.
}
\keyword{datasets}
