% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mica}
\alias{mica}
\title{Sample of Camtrap DP formatted data}
\format{
An object of class \code{list} of length 26.
}
\source{
\url{https://camtrap-dp.tdwg.org/example/}
}
\usage{
mica
}
\description{
A \href{https://camtrap-dp.tdwg.org}{Camera Trap Data Package} as read by
\code{\link[=read_camtrap_dp]{read_camtrap_dp()}}.
The source data are derived from the \href{https://camtrap-dp.tdwg.org/example/}{Camtrap DP example dataset }.
}
\examples{
\dontrun{
# mica.rda was created with:
mica <- read_camtrap_dp(
  "https://raw.githubusercontent.com/tdwg/camtrap-dp/b5c32179a9f072781386ca02e255e442a70902fa/example/datapackage.json"
)
save(mica, file = "data/mica.rda")
}
}
\seealso{
Other sample data: 
\code{\link{animal_positions}},
\code{\link{dep_calib_models}}
}
\concept{sample data}
\keyword{datasets}
