% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_species.R
\name{n_species}
\alias{n_species}
\title{Get number of identified species for each deployment}
\usage{
n_species(x)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{deploymentID}: Deployment unique identifier.
\item \code{n}: Number of observed and identified species.
}
}
\description{
Gets the number of identified species per deployment.
}
\examples{
x <- example_dataset()

# Get number of species
n_species(x)
}
\seealso{
Other exploration functions: 
\code{\link{camtrapR_cameraOperation}()},
\code{\link{camtrapR_recordTable}()},
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{get_scientific_name}()},
\code{\link{get_species}()}
}
\concept{exploration functions}
