% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{observations<-}
\alias{observations<-}
\title{Get or set observations}
\usage{
observations(x) <- value
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}

\item{value}{A data frame to assign as observations.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with observations.
}
\description{
Re-export of \code{\link[camtrapdp:observations]{camtrapdp::observations<-()}}.
}
\examples{
x <- example_dataset()
# Get the observations
observations(x)

# Set observations (not recommended outside a function)
observations(x) <- head(observations(x), 1)
}
\seealso{
Other accessor functions: 
\code{\link{add_coordinates}()},
\code{\link{deployments}()},
\code{\link{deployments<-}()},
\code{\link{events}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{media<-}()},
\code{\link{observations}()},
\code{\link{taxa}()}
}
\concept{accessor functions}
