% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\title{Retrieve observations data from a Camera Trap Data Package}
\usage{
observations(package)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{read_camtrap_dp()}.}
}
\value{
A tibble with the observations of the \code{package}
}
\description{
Retrieve observations data from a Camera Trap Data Package
}
\examples{
observations(mica)
}
