% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicates.R
\name{pred_primitive}
\alias{pred_primitive}
\title{Primitive filter predicate constructor}
\usage{
pred_primitive(arg, value, symbol, type)
}
\arguments{
\item{arg}{a character with the argument of the filter predicate}

\item{value}{the value the filter predicate uses to value \code{arg}. It can be a
number, a character, a date object or a POSIXct object}

\item{symbol}{a character with the symbol relation of the filter predicate}

\item{type}{a character with the type of the filter predicate}
}
\value{
a filter predicate object
}
\description{
This function is a primitive function to build all basic one arg - one value
filter predicates
}
\examples{
\dontrun{
pred_primitive(arg = "a", value = 5, symbol = ">", type = "greaterThan")
}
}
\keyword{internal}
