% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pos.R
\name{predict_pos}
\alias{predict_pos}
\title{Predict position}
\usage{
predict_pos(dat, mods, dep_tag = "deployment", x = "x", y = "y")
}
\arguments{
\item{dat}{Dataframe of animal position digitisation data.
It must contain (at least) the following columns:
\itemize{
\item columns defined in args \code{x}, \code{y} and \code{dep_tag}.
\item \code{ImageWidth}, \code{ImageHeight}: x and y pixel dimensions of each image. It
must be consistent for each deployment.
}}

\item{mods}{Named list of deployment calibration models.}

\item{dep_tag}{Character naming the column within \code{dat} against which names of the
elements can be matched to apply the right deployment calibration models.
Default: \code{"deployment"}.}

\item{x}{Character naming the column within \code{dat} containing x pixel
positions for each digitised point. Default: \code{"x"}.}

\item{y}{Character naming the column within \code{dat} containing y pixel
positions for each digitised point. Default: \code{"y"}.}
}
\value{
Original (tibble) dataframe as passed via \code{dat} with additional
columns:
\itemize{
\item \code{radius}: radial distance from camera
\item \code{angle}: angular distance from camera
\item \code{frame_count}: indicator of the frame order within each sequence
}
}
\description{
Predicts position relative to camera given image pixel positions and site
calibration models.
}
\examples{
To be done
}
