% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camtrap_dp.R
\name{read_camtrap_dp}
\alias{read_camtrap_dp}
\title{Read a Camtrap DP}
\usage{
read_camtrap_dp(file = NULL, media = TRUE, path = lifecycle::deprecated())
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}

\item{media}{If \code{TRUE} (default), read media records into memory. If \code{FALSE},
ignore media file to speed up reading larger Camtrap DP packages.}

\item{path}{Path to the directory containing the datapackage. Use  \code{file}
with path or URL to a \code{datapackage.json} file instead.}
}
\value{
List describing a Data Package (as returned by
\code{\link[frictionless:read_package]{frictionless::read_package()}}) containing the original metadata, as well
as a property \code{data} containing the data as three data frames:
\enumerate{
\item \code{deployments}
\item \code{media}
\item \code{observations}
}
}
\description{
Reads files from a \href{https://tdwg.github.io/camtrap-dp}{Camera Trap Data Package} into memory.
All datetime information is automatically transformed to Coordinated
Universal Time (UTC).
Vernacular names found in the metadata (\code{package$taxonomic}) are added to the
\code{observations} data frame.
}
\examples{
\dontrun{
# Provide path to a dataset descriptor file (i.e. a datapackage.json)
descriptor_file <- "https://raw.githubusercontent.com/tdwg/camtrap-dp/main/example/datapackage.json"

# Read Camtrap DP package into memory
dataset <- read_camtrap_dp(descriptor_file)

# The function will warn if it encounters parsing issues reading the data
# Those can be expected with readr::problems()
readr::problems(dataset$data$deployments)
readr::problems(dataset$data$media)
readr::problems(dataset$data$observations)
}
}
\seealso{
Other read functions: 
\code{\link{read_wi}()}
}
\concept{read functions}
