% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camtrap-dp.R
\name{read_camtrap_dp}
\alias{read_camtrap_dp}
\title{Read camtrap-dp formatted data}
\usage{
read_camtrap_dp(path)
}
\arguments{
\item{path}{a path to the folder containing camera trap files}
}
\value{
a list of tibbles (dataframes)
}
\description{
This function reads cameratrap data formatted following the \href{https://github.com/tdwg/camtrap-dp}{Camera Trap Data Package} (\href{https://github.com/tdwg/camtrap-dp}{Camtrap DP}) format.
}
\examples{
\dontrun{
library(here)
camtrap_dp_folder <- here("inst", "extdata", "gmu8-monitoring-faunabeheerzone-8-20210301093537")
gmu8 <- read_camtrap_dp(camtrap_dp_folder)
}
}
