% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camtrap_dp.R
\name{read_camtrap_dp}
\alias{read_camtrap_dp}
\title{Read camtrap-dp formatted data}
\usage{
read_camtrap_dp(file = NULL, media = TRUE, path = lifecycle::deprecated())
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}

\item{media}{If \code{TRUE}, read media records into memory. If \code{FALSE}, ignore
media file to speed up reading larger Camtrap DP packages.}

\item{path}{Path to the directory containing the datapackage. Use  \code{file}
with path or URL to a \code{datapackage.json} file instead.}
}
\value{
A list containing three (tibble) data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{media}
}

and a list with metadata: \code{datapackage}.
}
\description{
This function reads camera trap data formatted following the \href{https://github.com/tdwg/camtrap-dp}{Camera Trap Data Package (Camtrap DP)} format. The
function is built upon the functions \link[frictionless]{read_package} and
\link[frictionless]{read_resource}. This means a.o. that all datetime
information included in the camera trap data package is automatically
transformed to UTC (Coordinated Universal Time).
}
\details{
Vernacular names are typically used while working with camera trap
\emph{observations}, so they are added to the observations as defined in the
metadata (slot \code{taxonomic}), if present.
}
\examples{
\dontrun{
# Read Camtrap DP package
camtrap_dp_file <-  system.file("extdata", "mica", "datapackage.json", package = "camtraptor")
muskrat_coypu <- read_camtrap_dp(camtrap_dp_file)

# Read Camtrap DP package and ignore media file
muskrat_coypu <- read_camtrap_dp(camtrap_dp_file, media = FALSE)

# If parsing issues while reading deployments, observations or media arise,
use readr::problems()
camtrap_dp_file_with_issues <- system.file(
  "extdata",
  "mica_parsing_issues",
  "datapackage_for_parsing_issues.json",
  package = "camtraptor"
)
muskrat_coypu_with_issues <- read_camtrap_dp(camtrap_dp_file_with_issues, media = FALSE)
readr::problems(muskrat_coypu_with_issues$deployments)
readr::problems(muskrat_coypu_with_issues$observations)
readr::problems(muskrat_coypu_with_issues$media)
}
}
