% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camtrap_dp.R
\name{read_camtrap_dp}
\alias{read_camtrap_dp}
\title{Read camtrap-dp formatted data}
\usage{
read_camtrap_dp(file, media = TRUE, path = lifecycle::deprecated())
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}

\item{media}{If \code{TRUE}, read media records into memory. If \code{FALSE},
ignore media file to speed up reading larger Camtrap DP packages.}
}
\value{
A list containing three (tibble) data.frames:
\enumerate{
\item \code{observations}
\item \code{deployments}
\item \code{media}
}

and a list with metadata: \code{datapackage}
}
\description{
This function reads camera trap data formatted following the \href{https://github.com/tdwg/camtrap-dp}{Camera Trap Data Package (Camtrap DP)} format.
}
\details{
Vernacular names are typically used while working with camera trap
\emph{observations}, so they are added to the observations as defined in the
metadata (slot \code{taxonomic}), if present.
}
\examples{
\dontrun{
library(here)
# Read Camtrap DP package
camtrap_dp_dir <- here("inst", "extdata", "mica", "datapackage.json")
muskrat_coypu <- read_camtrap_dp(camtrap_dp_dir)

# Read Camtrap DP package and ignore media file
muskrat_coypu <- read_camtrap_dp(camtrap_dp_dir, media = FALSE)
}
}
