% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camtrap_dp.R
\name{read_camtrap_dp}
\alias{read_camtrap_dp}
\title{Read a Camera Trap Data Package}
\usage{
read_camtrap_dp(file)
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}
}
\value{
Camera Trap Data Package object.
}
\description{
Reads files from a \href{https://camtrap-dp.tdwg.org}{Camera Trap Data Package (Camtrap DP)} into memory.
}
