% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wi.R
\name{read_wi}
\alias{read_wi}
\title{Import Wildlife Insight camera trap data to Camera Trap Data Package}
\usage{
read_wi(
  directory = ".",
  rightsHolder = "project_admin_organization",
  coordinateUncertaintyInMeters = 30,
  captureMethod = "motion detection"
)
}
\arguments{
\item{directory}{Path to local directory to read the WI files}

\item{rightsHolder}{Acronym of the organization owning or managing the
rights over the data.}

\item{coordinateUncertaintyInMeters}{Uncertainty of the coordinate in meters.}

\item{captureMethod}{Either \code{"motion detection"} or \code{"time lapse"}.}
}
\value{
CSV (data) files written to disk.
}
\description{
Transforms a \href{https://www.wildlifeinsights.org/}{Wildlife Insight (WI)}
dataset to \href{https://tdwg.github.io/camtrap-dp/}{Camera Trap Data Package}
data exchange format.
}
\details{
Wildlife Insight offers to export your dataset \href{https://www.wildlifeinsights.org/get-started/download/private}{as private download} or
\href{https://www.wildlifeinsights.org/get-started/data-download/public}{public download}.
This function read the zip file from the download and convert the Projects,
Cameras, Deployments csv file to Darwin Core standard
}
