% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wi.R
\name{read_wi}
\alias{read_wi}
\title{Read a Wildlife Insights export}
\usage{
read_wi(directory = ".")
}
\arguments{
\item{directory}{Path to local directory to read files from.
The function expects \code{projects.csv}, \code{deployments.csv}, \code{cameras.csv}, and
\code{images.csv}.}
}
\value{
CSV (data) files written to disk.
}
\description{
Reads files from an unzipped \href{https://www.wildlifeinsights.org/}{Wildlife Insights} export into memory.
Data can be exported from Wildlife Insights as a \href{https://www.wildlifeinsights.org/get-started/data-download/public}{public} or
\href{https://www.wildlifeinsights.org/get-started/download/private}{private}
download.
The function transforms data and metadata to a \href{https://tdwg.github.io/camtrap-dp}{Camera Trap Data Package} which can be written to file with
\code{\link[frictionless:write_package]{frictionless::write_package()}}.
}
\details{
\strong{The function has only been tested on image-based projects.}
}
\seealso{
Other read functions: 
\code{\link{read_camtrap_dp}()}
}
\concept{read functions}
