% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_camtrapdp}
\title{Objects exported from other packages}
\seealso{
Other read functions: 
\code{\link{read_wi}()}
}
\concept{read functions}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{camtrapdp}{\code{\link[camtrapdp]{read_camtrapdp}}}
}}

