% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_camtrapdp}
\alias{filter_deployments}
\alias{filter_media}
\alias{filter_observations}
\title{Objects exported from other packages}
\seealso{
Other read functions: 
\code{\link{read_wi}()}

Other filter functions: 
\code{\link{apply_filter_predicate}()},
\code{\link{pred}()}

Other filter functions: 
\code{\link{apply_filter_predicate}()},
\code{\link{pred}()}

Other filter functions: 
\code{\link{apply_filter_predicate}()},
\code{\link{pred}()}
}
\concept{filter functions}
\concept{read functions}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{camtrapdp}{\code{\link[camtrapdp]{filter_deployments}}, \code{\link[camtrapdp]{filter_media}}, \code{\link[camtrapdp]{filter_observations}}, \code{\link[camtrapdp]{read_camtrapdp}}}
}}

