% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_deployments.R,
%   R/summarize_observations.R
\name{summarize_deployments}
\alias{summarize_deployments}
\alias{summarise_deployments}
\alias{summarise_observations}
\title{Summarize deployments information}
\usage{
summarize_deployments(x, group_by = "deploymentID", group_time_by = NULL)

summarise_deployments(x, group_by = "deploymentID", group_time_by = NULL)

summarise_observations(x, group_by = c("deploymentID"), group_time_by = NULL)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{group_by}{Character vector with the names of the columns in
deployments. At the moment you can choose one or many columns among:
\code{c("deploymentID", "locationID", "locationName", "deploymentTags")}.
Default: \code{deploymentID}.}

\item{group_time_by}{Character, one of \code{"day"}, \code{"week"}, \code{"month"},
\code{"year"}. The effort is calculated at the interval rate defined in
\code{group_time_by}. Default: \code{NULL}, no grouping, i.e. the entire duration of
the deployment is taken into account as a whole.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{group_by} names, e.g. \code{deploymentID} and \code{locationName}.
\item \code{group_time_by} name if provided, e.g. \code{month}. It contains the first
date of the time interval, e.g. the first day of the month.
\item \code{effort_duration}: A duration object (duration is a class from lubridate
package). Duration is always recorded as a fixed number of seconds. See
\code{\link[lubridate:duration]{lubridate::duration()}}.
}
}
\description{
Summarizes deployments information, more specifically the duration effort.
}
\details{
\code{summarize_deployments()} and \code{summarise_deployments()} are synonyms.
}
\examples{
x <- example_dataset()

# Return effort per `deploymentID` and `locationName`, by day
summarize_deployments(
  x,
  group_by = c("deploymentID", "locationName"),
  group_time_by = "day"
)

# Return effort per `deploymentID` and `locationName`. No grouping by time.
summarize_deployments(
  x,
  group_by = c("deploymentID", "locationName")
)

# Afterwards, you can calculate the total effort over all deployments. You
can also show other information, e.g. the (number of) deployments and
locations.
library(dplyr)
summarize_deployments(
  x,
  group_by = c("deploymentID", "locationName"),
  group_time_by = "month"
  ) \%>\%
dplyr::group_by(month) \%>\%
dplyr::summarise(
  deploymentIDs = list(deploymentID),
  ndep = length(unique(deploymentID)),
  nloc = length(unique(locationName)),
  effort_duration = sum(effort_duration)
)
}
\seealso{
Other exploration functions: 
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{n_species}()},
\code{\link{summarize_observations}()}
}
\concept{exploration functions}
