% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_deployments.R
\name{summarize_deployments}
\alias{summarize_deployments}
\title{Summarize deployments information}
\usage{
summarize_deployments(x, group_by = "deploymentID", group_time_by = NULL)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{group_by}{Character vector with the names of the columns in
deployments. At the moment you can choose one or many columns among:
\code{c("deploymentID", "locationID", "locationName", "deploymentTags")}.
Default: \code{deploymentID}.}

\item{group_time_by}{Character, one of \code{"day"}, \code{"week"}, \code{"month"},
\code{"year"}. The effort is calculated at the interval rate defined in
\code{group_time_by}. Default: NULL, no grouping, i.e. the entire duration of
the deployment is taken into account as a whole.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{group_by} names, e.g. \code{deploymentID} and \code{locationName}.
\item \code{group_time_by} name if provided, e.g. \code{month}. It contains the first
date of the time interval, e.g. the first day of the month.
\item \code{effort_duration}: A duration object (duration is a class from lubridate
package). Duration is always recorded as a fixed number of seconds. See
\code{\link[lubridate:duration]{lubridate::duration()}}.
}
}
\description{
This function is intended to summarize deployments information, such as the
effort.
}
\examples{
x <- example_dataset()

# Return effort per deploymentID and locationName, by day
summarize_deployments(
  x,
  group_by = c("deploymentID", "locationName"),
  group_time_by = "day"
)

# Return effort per deploymentID and locationName, by month
summarize_deployments(
 x,
 group_by = c("deploymentID", "locationName"),
 group_time_by = "month"
)

# Return the effort per deploymentID and locationName. No grouping by time.
summarize_deployments(
  x,
  group_by = c("deploymentID", "locationName")
)
}
\seealso{
Other exploration functions: 
\code{\link{get_custom_effort}()},
\code{\link{get_effort}()},
\code{\link{get_n_individuals}()},
\code{\link{get_n_obs}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{n_species}()}
}
\concept{exploration functions}
