% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_observations.R
\name{summarize_observations}
\alias{summarize_observations}
\title{Summarize observations information}
\usage{
summarize_observations(
  x,
  group_by = c("deploymentID", "scientificName"),
  group_time_by = NULL
)
}
\arguments{
\item{x}{Camera trap data package object, as returned by
\code{\link[camtrapdp:read_camtrapdp]{camtrapdp::read_camtrapdp()}}.}

\item{group_by}{Character vector with names of columns in deployments and
observations. At the moment you can choose one or many columns among:
\code{c("deploymentID", "locationID", "locationName", "deploymentTags", "scientificName", "lifeStage", "sex", "behavior")}. Default:
\code{c("deploymentID", "scientificName")}.}

\item{group_time_by}{Character, one of \code{"day"}, \code{"week"}, \code{"month"},
\code{"year"}. The features are calculated at the interval rate defined in
\code{group_time_by}. Default: \code{NULL}, no grouping, i.e. the entire duration of
the deployment is taken into account as a whole.}
}
\value{
A tibble data frame with the following columns:
\itemize{
\item \code{group_by} names, e.g. \code{deploymentID} and \code{locationName}.
\item \code{group_time_by} name if provided, e.g. \code{month}. It is a datetime column
containing the first date of the time interval, e.g. the first day of the
month.
\item \code{n_observations}: integer vector with the number of observations.
\item \code{sum_count}: integer vector with the sum of individual counts.
}
}
\description{
Summarizes event-based observations by calculating:
\itemize{
\item Number of observations.
\item Sum of individual counts.
\item Number of species.
\item Relative Abundance Index (RAI) based on number of observations.
\item Relative Abundance Index (RAI) based on individual counts.
}
}
\details{
\code{summarize_observations()} and \code{summarise_observations()} are synonyms.
}
\examples{
x <- example_dataset()
# Summarize observations by `deploymentID` and `scientificName` (default)
summarize_observations(x)

# Summarize observations by `deploymentID` and month
summarize_observations(x, group_time_by = "month")

# Summarize observations by `locationId`, and `locationName`
#' summarize_observations(x, group_by = "locationName")
}
\seealso{
Other exploration functions: 
\code{\link{get_n_individuals}()},
\code{\link{get_rai}()},
\code{\link{get_rai_individuals}()},
\code{\link{n_species}()},
\code{\link{summarize_deployments}()}
}
\concept{exploration functions}
