% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapdp-reexports.R
\name{taxa}
\alias{taxa}
\title{Get taxa}
\usage{
taxa(x)
}
\arguments{
\item{x}{Camera Trap Data Package object, as returned by \code{\link[camtrapdp:read_camtrapdp]{read_camtrapdp()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with the taxonomic information,
containing at least a \code{scientificName} column.
}
\description{
Re-export of \code{\link[camtrapdp:taxa]{camtrapdp::taxa()}}.
}
\examples{
x <- example_dataset()
taxa(x)
}
\seealso{
Other accessor functions: 
\code{\link{add_coordinates}()},
\code{\link{contributors}()},
\code{\link{deployments}()},
\code{\link{events}()},
\code{\link{individuals}()},
\code{\link{locations}()},
\code{\link{media}()},
\code{\link{observations}()}
}
\concept{accessor functions}
