% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_species_per_deployment.R
\name{visualize_species_per_deployment}
\alias{visualize_species_per_deployment}
\title{Visualize number of species per deployment}
\usage{
visualize_species_per_deployment(
  deployments,
  observations,
  cluster = TRUE,
  hover_column = "n",
  relative_color_scale = TRUE,
  max_color_scale = NULL
)
}
\arguments{
\item{deployments}{a tibble (data.frame) containing deployments}

\item{observations}{a tibble (data.frame) containing observations}

\item{cluster}{a logical value indicating whether using the cluster option
while visualizing maps. Default: TRUE}

\item{hover_column}{character with the name of the column to use for showing
location deployment information while hovering the mouse over. One from: -
\code{n}: number of species (default) - \code{deployment_id} - \code{location_id} -
\code{location_name} - \code{NULL}: hovering disabled}

\item{relative_color_scale}{a logical indicating whether to use a relative
color scale (\code{TRUE}) or an absolute scale (\code{FALSE}). If absolute scale is
used, specify a valid \code{max_color_scale}}

\item{max_color_scale}{a number indicating the max value used in the color
scale}
}
\value{
a leaflet map
}
\description{
This function visualizes the number of species detected by each deployment.
The circle size and color are proportional to the number of detected species.
Deployments without observations are shown as gray circles and a message is
returned.See \href{https://github.com/inbo/camtrapdp/issues/6}{issue #6}.
Unidentified species (empty scientific name) count as 0.
}
\details{
Possible filtering about time period, ... (to be added)
}
\examples{
\dontrun{
# cluster enabled and  number of species (`n`) shown while hovering
visualize_species_per_deployment(camtrapdp$deployments,
                                 camtrapdp$observations)
# cluster disabled
visualize_species_per_deployment(camtrapdp$deployments,
                                 camtrapdp$observations,
                                 cluster = FALSE)
# show location name while hovering
visualize_species_per_deployment(camtrapdp$deployments,
                                 camtrapdp$observations,
                                 hover_column = "location_name")

# use absolute scale for colors
visualize_species_per_deployment(camtrapdp$deployments,
                                 camtrapdp$observations,
                                 relative_color_scale = FALSE,
                                 max_color_scale = 4)
}
}
