% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_species_per_deployment.R
\name{visualize_species_per_deployment}
\alias{visualize_species_per_deployment}
\title{Visualize number of species per deployment}
\usage{
visualize_species_per_deployment(
  deployments,
  observations,
  cluster = TRUE,
  hover_column = "deployment_id"
)
}
\arguments{
\item{deployments}{a tibble (data.frame) containing deployments}

\item{observations}{a tibble (data.frame) containing observations}

\item{cluster}{a logical value indicating whether using the cluster option
while visualizing maps. Default: TRUE}

\item{hover_column}{character with the name of the column to use for showing
location deployment information while hovering the mouse over. By default:
\code{deployment_id}. Use \code{NULL} to disable hovering.}
}
\value{
a leaflet map
}
\description{
This function visualizes the number of species detected by each deployment.
Possible filtering about time period, ... (to be added)
}
\examples{
\dontrun{
# cluster enabled and  `deployment_id` column for hovering
visualize_species_per_deployment(camtrapdp$deployments,
                                 camtrapdp$observations)
# cluster disabled
visualize_species_per_deployment(camtrapdp$deployments,
                                 camtrapdp$observations,
                                 cluster = FALSE)
visualize_species_per_deployment(camtrapdp$deployments,
                                 camtrapdp$observations,
                                 hover_column = "location_name")
}
}
