% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform Camtrap DP data to Darwin Core}
\usage{
write_dwc(package, directory = ".")
}
\arguments{
\item{package}{A Camtrap DP, as read by \code{\link[=read_camtrap_dp]{read_camtrap_dp()}}.}

\item{directory}{Path to local directory to write file to.}
}
\value{
CSV files written to disk.
}
\description{
Transforms the data of a \href{https://tdwg.github.io/camtrap-dp/}{Camera Trap Data Package (Camtrap DP)} to
\href{https://dwc.tdwg.org/}{Darwin Core} CSV files that can be uploaded to a
\href{https://www.gbif.org/ipt}{GBIF IPT} for publication.
A \code{meta.xml} file is not created.
}
\section{Transformation details}{

Data are transformed following best practices (Reyserhove et al. in prep.)
into an
\href{https://rs.gbif.org/core/dwc_occurrence_2022-02-02.xml}{Occurrence core} and
\href{https://rs.gbif.org/extension/ac/audubon_2020_10_06.xml}{Audubon Media Description extension}.
See the \link{SQL files}((https://github.com/inbo/camtraptor/tree/main/inst/sql)
used by this function for details.

The following terms are set from the \code{package} metadata:
\itemize{
\item \strong{datasetName}: Title as provided in \code{package$title}.
\item \strong{datasetID}: Identifier as provided in \code{package$id}.
Can be a DOI.
\item \strong{rightsHolder}: Rights holder as provided in \code{package$rightsHolder}.
\item \strong{collectionCode}: Platform name as provided in \code{package$platform$title}.
\item \strong{license}: License with scope \code{data} as provided in \code{package$licenses}.
\item \strong{rights} for media files: License with scope \code{media} as provided in
\code{package$licenses}.
}
}

\seealso{
Other export functions: 
\code{\link{write_eml}()}
}
\concept{export functions}
