% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwc.R
\name{write_dwc}
\alias{write_dwc}
\title{Transform Camtrap DP data to Darwin Core}
\usage{
write_dwc(package, directory = ".")
}
\arguments{
\item{package}{A Camtrap DP, as read by \code{\link[=read_camtrap_dp]{read_camtrap_dp()}}.}

\item{directory}{Path to local directory to write file(s) to.}
}
\value{
CSV file(s) written to disk.
}
\description{
Transforms the data of a \href{https://tdwg.github.io/camtrap-dp/}{Camera Trap Data Package} to \href{https://dwc.tdwg.org/}{Darwin Core}
CSV file(s) that can be uploaded to a \href{https://www.gbif.org/ipt}{GBIF IPT} for
publication.
A \code{meta.xml} file is not created.
}
\section{Transformation details}{

Data are transformed into an
\href{https://rs.gbif.org/core/dwc_occurrence_2022-02-02.xml}{Occurrence core} and
\href{https://rs.gbif.org/extension/ac/audubon_2020_10_06.xml}{Audubon Media Description extension}.
This \strong{follows recommendations} discussed and created by Peter Desmet,
John Wieczorek, Lien Reyserhove, Ben Norton and others.
See the \href{https://github.com/inbo/camtraptor/tree/main/inst/sql}{SQL file(s)}
used by this function for details.

The following terms are set from the \code{package} metadata:
\itemize{
\item \strong{datasetName}: Title as provided in \code{package$title}.
\item \strong{datasetID}: Identifier as provided in \code{package$id}.
Can be a DOI.
\item \strong{rightsHolder}: Rights holder as provided in \code{package$rightsHolder}.
\item \strong{collectionCode}: Platform name as provided in \code{package$platform$title}.
\item \strong{license}: License with scope \code{data} as provided in \code{package$licenses}.
\item \strong{rights} for media files: License with scope \code{media} as provided in
\code{package$licenses}.
\item \strong{dwc:dataGeneralizations}: "coordinates rounded to
\code{package$coordinatePrecision} degrees".
\item \strong{coordinatePrecision}: \code{package$coordinatePrecision} (e.g. \code{0.001}).
}

Key features of the Darwin Core transformation:
\itemize{
\item Deployments (of camera traps) are parent events, with observations
(machine observations) as child events. No information about the parent
event is provided other than its ID, meaning that data can be expressed in
an Occurrence Core with one row per observation and \code{parentEventID} shared
by all occurrences in a deployment.
\item Sequence-based observations share an \code{eventID} per sequence, image-based
observations share an \code{eventID} per image.
\item The image(s) an observation is based on are provided in the \href{https://rs.gbif.org/extension/ac/audubon_2020_10_06.xml}{Audubon Media Description extension}, with a foreign
key to the observation.
\item Excluded are records that document blank or unclassified media, vehicles
and observations of humans.
}
}

\seealso{
Other publication functions: 
\code{\link{round_coordinates}()},
\code{\link{write_eml}()}
}
\concept{publication functions}
